### 2020-05-23 #2234 @whedon
[PRE REVIEW]: WDPM: the Wetland DEM Ponding Model
**Submitting author:** @KevinShook (<a href="http://orcid.org/0000-0003-0363-9929">Kevin Shook</a>)
**Repository:** <a href="https://github.com/CentreForHydrology/WDPM" target ="_blank">https://github.com/CentreForHydrology/WDPM</a>
**Version:** v1.0.0
