+++ paper.bib
@@ -17,7 +17,7 @@
   volume = {2667},
   pages = {2650--2667},
   issn = {08856087},
-  doi = {10.1002/hyp},
+  doi = {10.1002/hyp.6787},
   file = {/home/kevin/data/zotero/storage/PCHIC3ZJ/Pomeroy et al. - 2007 - The cold regions hydrological model a platform for basing process representation and model structure on physica.pdf},
   journal = {Hydrological Processes},
   keywords = {accepted 12 december 2006,blowing snow,distributed modelling,evaporation,frozen soil infiltration,hillslope,hydrological modelling,hydrological response unit,physically based modelling,received 1 june 2006,runoff,snowmelt,subsurface flow},
@@ -87,7 +87,7 @@
   journal = {Journal of Hydrology}
 }
 
-@article{spence:2003,
+@article{spenceHydrologySubarcticCanadian2003,
   title = {Hydrology of Subarctic {{Canadian}} Shield: {{Soil}}-Filled Valleys},
   author = {Spence, Christopher and Woo, Ming Ko},
   year = {2003},
