+++ paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-The hydrography of the Prairies of western North America, including Canada, is unusual in that the landscape is flat and recently formed due to the effects of pleistocene glaciation and a semi-arid climate since holocene deglaciation. Therefore, there has not been sufficient energy, time, or runoff water to carve typical dendritic surface water drainage networks in many locations. In these regions, runoff is often detended and sometimes stored by the millions of depressions (known locally as "potholes" or "sloughs") that cover the landscape. Conventional hydrological models are unable to simulate the spatial distribution of ponded water in Prairie basins dominated by depressional storage. When the depressions are filled, the detended water may overflow to another depression, though a process known as "fill and spill" [@spence:2003]. Therefore, the fraction of a depression-dominated Prairie basin that contributes flow to the outlet will change dynamically with the state of water storage within the basin. This situation is difficult to simulate with conventional hydrological models. 
+The hydrography of the Prairies of western North America, including Canada, is unusual in that the landscape is flat and recently formed due to the effects of pleistocene glaciation and a semi-arid climate since holocene deglaciation. Therefore, there has not been sufficient energy, time, or runoff water to carve typical dendritic surface water drainage networks in many locations. In these regions, runoff is often detended and sometimes stored by the millions of depressions (known locally as "potholes" or "sloughs") that cover the landscape. Conventional hydrological models are unable to simulate the spatial distribution of ponded water in Prairie basins dominated by depressional storage. When the depressions are filled, the detended water may overflow to another depression, though a process known as "fill and spill" [@spenceHydrologySubarcticCanadian2003]. Therefore, the fraction of a depression-dominated Prairie basin that contributes flow to the outlet will change dynamically with the state of water storage within the basin. This situation is difficult to simulate with conventional hydrological models. 
 
 # Program description
 
