### 2020-05-23 a538e3f
fixed figure references
