+++ paper.md
@@ -68,7 +68,7 @@ Following the addition and draining of water, 200 mm of water was removed using 
 
 
 The WDPM is very computationally expensive, requiring many thousands or millions of iterations, particularly to drain water from the DEM (the addition and subtraction of water being less expensive). 
-In the examples above, the addition of 300 mm of water, smoothed to a resolution of 100 mm, required 700 iterations. Draining the water to a tolerance of 0.1 mm/0.1 m^3^ required 332,000 iterations and took 981 s on a system with an Intel i7 (4 cores, 8 threads) running Linux. Removing the water and smoothing to a tolerance of 1 mm required 1,000 iterations - the minimum possible. 
+In the examples above, the addition of 300 mm of water, smoothed to a resolution of 100 mm, required 7000 iterations. Draining the water to a tolerance of 0.1 mm/0.1 m^3^ required 332,000 iterations and took 981 s on a system with an Intel i7 (4 cores, 8 threads) running Linux. Removing the water and smoothing to a tolerance of 1 mm required 1,000 iterations - the minimum possible. 
 
 Despite the computational cost, the program's great advantage is that it can simulate any storage state within a complex system of Prairie depressions.  This capability makes it useful for mapping the flooplains from non-riverine floods in the Canadian Prairies. The program outputs have been verified by remote sensing of recent floods in this region, and WDPM has been used to develop a simpler parametric model that is more easily incorporated in hydrological models [@shookStorageDynamicsSimulations2013]. The program's floodplain mapping capabiltiies
 have made it useful for operational flood hazard mapping [@armstrongUSINGWETLANDPONDING2013a] in the Prairies by several government agencies and by private consultants.
