### 2020-05-28 #2234 @kbarnhart
@openjournals/joss-eics I can handle this one.
