### 2020-05-28 #2234 @kbarnhart
@KevinShook thanks for submitting this to JOSS. If you have any recommendations for reviewers, please post them here (separate @ from handle name, e.g., you would refer to me as @ kbarnhart). Consider the sheet with potential reviewers [linked above](https://bit.ly/joss-reviewers), but don't feel limited to it. It is OK if potential reviewers are not yet on GitHub. I have some initial thoughts for suitable reviewers but am also interested in your input.
