### 2020-05-29 #2234 @KevinShook
@kbarnhart thanks very much for getting in touch.
I'm afraid that I can't think of anyone who would be good as a reviewer, as everyone I know is either a hydrologist, or a computer scientist (or someone I work with!). Few people straddle both disciplines.
Looking at the list of reviewers, I see that there are about 12 who are comfortable with hydrology, of whom 4 (mlep, dvalters, arbennett, sgrieve) mention C as a preferred or other language. Although the GUI of WDPM was written in Python, it is not necessary to execute the program, so is probably less important.
I see that dvalters is familiar with C, Python and parallel processing, which is very important to this program, so would seem to be a good selection.
I see that there are another 13 parallel processing people, but none of them mention OpenCl.
Did you have some other reviewers in mind as well?
