### 2020-06-07 CentreForHydrology/WDPM#1 @KCallaghan
'cmap_black.sh' file is missing
I am trying to run this using the WDPM.py file and it calls for a cmap_black.sh file, which is not included. The manual does not provide any description of this file for me to create it myself, so I suspect it should have been provided.

openjournals/joss-reviews#2276
