### 2020-06-07 CentreForHydrology/WDPM#1 @KevinShook
You are quite right. It's distributed with the binary version of the program.
I'll add it to the repository.
Thanks very much
