### 2020-06-07 CentreForHydrology/WDPM#1 @KCallaghan
Thanks.
I downloaded the compiled version as well, and I could get the GUI launched by copying over the cmap_black.sh file from there (but please do still add it here!)
This is minor, but the version I compiled myself works better since the buttons don't appear properly on the compiled version. I couldn't find where the dropdown was supposed to be for 'method' until I was able to open the GUI in the version I had compiled myself:

![image](https://user-images.githubusercontent.com/41647227/83959631-d13e8780-a844-11ea-996d-bb5ef7d2d79f.png)
