### 2020-06-07 b06c341
added missing files cmap_black.sh and colormap_black.txt
