+++ src/cmap_black.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+startingDir=`pwd`
+# creates color map from water applied to DEM
+infile=$1
+outfile="${infile%.*}"
+gdaldem color-relief $infile $startingDir/colormap_black.txt -OF png $outfile".png"
+# delete superfluous xml file
+rm $outfile".png.aux.xml"
