+++ src/colormap_black.txt
@@ -1,0 +1,4 @@
+3,25,0,230
+0.001,25,0,230
+0,yellow
+-9999, black
