### 2020-06-07 CentreForHydrology/WDPM#1 @KevinShook
Which OS are you using? I seem to remember the same problem with the compiled version under Linux - but I never use the GUI.
We'll update the compiled version.
