### 2020-06-07 CentreForHydrology/WDPM#1 @KCallaghan
I'm using Linux Mint.
I haven't figured out how to not use the GUI yet, but I'll get there!
