### 2020-06-07 CentreForHydrology/WDPM#2 @KCallaghan
Frequent GUI crashes
With the error message:

> Gdk-Message: 22:39:47.991: WDPM.py: Fatal IO error 11 (Resource temporarily unavailable) on X server :0.

This happens when I click 'start', but only sometimes with a given set of parameters.

openjournals/joss-reviews#2276
