### 2020-06-07 CentreForHydrology/WDPM#1 @KevinShook
That's great - I use Mint as well.
I'll check to see what's going on.
The parameters that you used look fine.
Thanks for checking all this
