### 2020-06-12 CentreForHydrology/WDPM#4 @KCallaghan
Installation of Linux binary gives errors
This is an issue with the downloadable Linux binary, not the GitHub source code, but this still seemed the best place to address it.
I realised that I had just downloaded and attempted to run the program and had missed some additional steps in the manual.
From the manual:
> Do sudo make all to build, install and test the Intel and Nvidia OpenCL drivers

When I attempt to do so, I get the following errors:

> ./installcl.sh: 18: ./installcl.sh: cannot create /etc/OpenCL/vendors/intel.icd: Directory nonexistent
> chmod: cannot access '/etc/OpenCL/vendors/intel.icd': No such file or directory
> Installation complete.
> gcc Deviceinfo.c -lOpenCL -fopenmp -o Deviceinfo -std=c99
> Deviceinfo.c: In function ‘CLInitDevices’:
> Deviceinfo.c:193:3: warning: ‘clCreateCommandQueue’ is deprecated [-Wdeprecated-declarations]
>    if( !(cl->queue = clCreateCommandQueue(cl->context, cl->currentDevice, 0, &result)) )
>    ^~
> In file included from /usr/include/CL/opencl.h:47:0,
>                  from Deviceinfo.c:9:
> /usr/include/CL/cl.h:1443:1: note: declared here
>  clCreateCommandQueue(cl_context                     /* context */,
>  ^~~~~~~~~~~~~~~~~~~~
> ./Deviceinfo
> Error is -1001
> Error: clGetPlatforms failed: Unknown.
>
> Makefile:8: recipe for target 'test' failed
> make: *** [test] Error 1

I am not sure if this could be leading to problems I am having with running this version of the program.

openjournals/joss-reviews#2276
