### 2020-06-16 #2276 @kbarnhart
@KevinShook, @r-barnes, @awickert, @KCallaghan

I wanted to check in and see how things were going with this review, and to answer any questions if you have them. As mentioned above, JOSS is presently requesting reviewers complete their reviews in six weeks (a bit more than three weeks from now).

@KCallaghan thanks for linking all of your issues to this one 👍 🎉
