### 2020-06-21 534a901
fixed GUI problems caused by change in Python 2
