### 2020-06-21 CentreForHydrology/WDPM#2 @KevinShook
The GUI issues appear to have been caused by a change in Python.
The code has not been touched in several years, and most of
the users never use the GUI, so it was missed.
Apparently there was a change from Python, which required
changing all of the instances of  `wx.OPEN`  to `wx.FD_OPEN`.
This has now been done, and the GUI now appears to run properly
