### 2020-06-21 CentreForHydrology/WDPM#1 @KevinShook
The file `cmap_black.sh` has been added, and the GUI Python code
has been changed to fix the problems caused by changes in Python
