### 2020-06-26 CentreForHydrology/WDPM#5 @r-barnes
Place paper and associated images in a separate directory
openjournals/joss-reviews#2276
