### 2020-06-26 CentreForHydrology/WDPM#6 @r-barnes
Remove .Rhistory and add to .gitignore
openjournals/joss-reviews#2276
