### 2020-06-26 CentreForHydrology/WDPM#7 @r-barnes
Executable files unavailabe
The link in the README (http://www.usask.ca/hydrology/WDPM.php) redirects to (https://research-groups.usask.ca/hydrology/). There are no executables available here.

Executables should rarely, if ever, be distributed with a Git repo. The repo should contain sufficient material to generate the executables.

openjournals/joss-reviews#2276
