### 2020-06-26 CentreForHydrology/WDPM#8 @r-barnes
imports should be sorted
Import should be sorted alphabetically and grouped, per the pep8 style guide ([link](https://www.python.org/dev/peps/pep-0008/#:~:text=Imports%20are%20always%20put%20at,Related%20third%20party%20imports.)).

> Imports are always put at the top of the file, just after any module comments and docstrings, and before module globals and constants.
>
> Imports should be grouped in the following order:
>
> Standard library imports.
> Related third party imports.
> Local application/library specific imports.
> You should put a blank line between each group of imports.

Problems include, but are not limited to [this](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L1-L10).

openjournals/joss-reviews#2276
