### 2020-06-26 CentreForHydrology/WDPM#9 @r-barnes
No need to support Python2
[These lines](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L13-L16), and probably others in the repo, are part of a Python2 compatibility strategy.

Python2 was [discontinued on January 1, 2020](https://www.python.org/doc/sunset-python-2/) after a 12 year decommissioning process. Supporting Python2 at this point is a Sisyphean task; it would be better to remove such support.

openjournals/joss-reviews#2276
