### 2020-06-26 CentreForHydrology/WDPM#10 @r-barnes
Remove `==True`
There are many instances of checking if something is true, e.g. [this](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L23). Doing so is non-Pythonic.

openjournals/joss-reviews#2276
