### 2020-06-26 CentreForHydrology/WDPM#11 @r-barnes
Remove large comment blocks
Production code should not have functionality commented out. If you're not using it, remove it. If you might use, make a flag of some sort to control access. [These lines](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L1058-L1285) are a notable instance.

openjournals/joss-reviews#2276
