### 2020-06-26 CentreForHydrology/WDPM#12 @r-barnes
Don't repeat yourself
[These blocks](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L1006-L1027) have near-duplicate functionality. Good coding practice is to abstract repetition into functions, please do so.

openjournals/joss-reviews#2276
