### 2020-06-26 CentreForHydrology/WDPM#13 @r-barnes
Add comments to UI generation
[This section](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L96-L228) is pretty dense. White space coupled with appropriate comments would make its maintenance easier.

openjournals/joss-reviews#2276
