### 2020-06-26 CentreForHydrology/WDPM#14 @r-barnes
Abstract UI control
[These lines](https://github.com/CentreForHydrology/WDPM/blob/master/src/WDPM.py#L294-L414) show a large amount of duplication. This could be reduced by looping over a list of components to enable/disable. It also suggests that perhaps the UI should be broken into separate controller classes that can be enabled/disabled separately. Handling everything in single mass code blocks is a recipe for low maintainability.

openjournals/joss-reviews#2276
