### 2020-06-26 CentreForHydrology/WDPM#7 @KevinShook
Unfortunately, the website was changed after the JOSS submission.
I can update the link in the README, unless you think that there should not be a link at all.
