### 2020-06-26 CentreForHydrology/WDPM#15 @r-barnes
What is `WDPM/src/WDPMCL`?
Mystery binary blobs should be avoided whenever possible: the repo itself should generating them for transparency, reproducibility, and mutability. I'm not sure what `WDPM/src/WDPMCL` is.

openjournals/joss-reviews#2276
