### 2020-06-26 CentreForHydrology/WDPM#16 @r-barnes
Need a build system
`WDPM/src/` contains C files and OpenCL files. These need a build system so they can be reproducibly compiled into a binary by the user. `CMake` is a strong, cross-platform choice. `Make` might be initially simpler to work with, but will be more challenging cross-platform.

An example of using CMake is [here](https://github.com/r-barnes/Barnes2019-DepressionHierarchy/blob/master/CMakeLists.txt). If you have difficulty working with cmake, please comment here and I can help.

openjournals/joss-reviews#2276
