### 2020-06-26 CentreForHydrology/WDPM#17 @r-barnes
Print statements need parens
Python2 was [discontinued](https://www.python.org/doc/sunset-python-2/) on January 1, 2020 after a 12 year decommissioning process.

At this point `print` statements need parens:
```python
print("This is now the only way to print in Python.")
```

openjournals/joss-reviews#2276
