### 2020-06-26 CentreForHydrology/WDPM#18 @r-barnes
Consider a linter
[Pylint](https://www.pylint.org/) and [Black](https://github.com/psf/black) are Python linters: programs that parse your code to detect mistakes, bugs, and bad practices before they become an issue. I use them regularly and many of my industry collaborators set up their git repos so they are physically unable to commit code that doesn't pass linting checks. Given the number of issues in the existing Python code, using a linter is probably the fast way to improve it.

I forget if Pylint or Black do this, but some linters complain if functions have too many `if` clauses or classes have too many methods. This can be difficult to resolve, but doing so would resolve some cases where I feel the existing logic is difficult to follow, especially since there are no comments.

openjournals/joss-reviews#2276
