### 2020-06-26 CentreForHydrology/WDPM#19 @r-barnes
Indicate dependencies
I haven't seen any file indicating code dependencies.

For Python, a [`requirements.txt`](https://pip.pypa.io/en/stable/reference/pip_install/#requirement-specifiers) file would be helpful.

For other dependencies, such as `gdaldem`, which appears in `WDPM/src/cmap_black.sh`, the README file or a conda requirements file should indicate the dependency.

openjournals/joss-reviews#2276
