### 2020-06-26 #2276 @r-barnes
That's it for me this morning; I'll circle back to inspecting the C and OpenCL code later.
