### 2020-06-26 CentreForHydrology/WDPM#7 @r-barnes
If the link is to binaries generated by the C and OpenCL programs in the repo, I think it would be better to focus your attention on the build system for the binaries. That eliminates the worry about maintaining the structure of the external website and ultimately make the work more reproducible. So, yes, I'd drop the link.

Let me know if I can be of assistance.
