### 2020-06-26 CentreForHydrology/WDPM#2 @r-barnes
If the GUI isn't widely used, you could reconsider whether it's worth having. I don't have a strong stance on this one way or the other.
