### 2020-06-26 CentreForHydrology/WDPM#16 @r-barnes
Actually, here's a decent `./CMakeLists.txt` file:
```
cmake_minimum_required (VERSION 3.9)

project (dephier LANGUAGES C)

find_package(OpenCL REQUIRED)

#Adds aggressive warnings to all compilation
if(MSVC)
  add_compile_options(/W4 /WX)
else()
  add_compile_options(-Wall -pedantic)
endif()

add_executable(wdpm-cl src/WDPMCL-Command-line.c)
target_link_libraries(wdpm-cl PRIVATE OpenCL)

add_executable(wdpm-gui src/WDPMCL-GUI.c)
target_link_libraries(wdpm-gui PRIVATE OpenCL)

#Include Python files in build output
ADD_CUSTOM_COMMAND(OUTPUT "${CMAKE_BINARY_DIR}/WDPM.py"
  COMMAND cmake -E copy_if_different "${CMAKE_CURRENT_SOURCE_DIR}/src/WDPM.py"
  "${CMAKE_BINARY_DIR}/WDPM.py"
  DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/src/WDPM.py"
  )
 list(APPEND pyfile_dest "${CMAKE_BINARY_DIR}/WDPM.py")
ADD_CUSTOM_TARGET(python ALL DEPENDS ${pyfile_dest})
```
See [this](https://github.com/r-barnes/Barnes2019-DepressionHierarchy) for an explanation of how to compile with `cmake`.

Everything, including the Python script, ends up in the `build` directory.
