### 2020-06-26 CentreForHydrology/WDPM#20 @r-barnes
Resolve C/OpenCL compilation warnings
Compiling with [this](https://github.com/CentreForHydrology/WDPM/issues/16#issuecomment-650282519) `CMakeLists.txt` file using the `-Wall -pedantic` combo reveals a number of code issues including unused variables, unchecked file I/O returns, and uninitialized variables. These should all be resolved.

openjournals/joss-reviews#2276
