### 2020-07-01 CentreForHydrology/WDPM#21 @KCallaghan
More detailed testing steps needed
The manual includes some examples, e.g. 100 mm of water added. However, it does not state what runoff fraction, elevation tolerance, etc was used to obtain the results seen so that a user can more easily test and compare results.

openjournals/joss-reviews#2276
