### 2020-07-01 CentreForHydrology/WDPM#22 @KCallaghan
Model execution times. 
The performance I am seeing is significantly slower than that listed in the manual. I tried testing using the supplied 'basin5.asc' file, adding 10 mm water with an elevation tolerance of 1 mm, running in serial.
I used a Lenovo Thinkpad T460p with i-7, 2.7 GHz x 4. A comparable system in the user guide took ~140 s to run this command. This took 1048 s for me.
I ran with:
> ./WDPMCL add basin5.asc null test.asc null 10.0 1.0 1.0 0 0

openjournals/joss-reviews#2276
