### 2020-07-01 CentreForHydrology/WDPM#23 @KCallaghan
Getting OpenCL working
I am having trouble running with OpenCL; I am not sure if the issue is with my installation of OpenCL or elsewhere. When I try running with OpenCL, either with CPU or GPU, I get a segmentation fault. I am able to run the program in serial.
In case it is helpful, 'clinfo' shows the following output:

> Number of platforms                               1
>   Platform Name                                   NVIDIA CUDA
>   Platform Vendor                                 NVIDIA Corporation
>   Platform Version                                OpenCL 1.2 CUDA 10.2.185
>   Platform Profile                                FULL_PROFILE
>   Platform Extensions                             cl_khr_global_int32_base_atomics cl_khr_global_int32_extended_atomics cl_khr_local_int32_base_atomics cl_khr_local_int32_extended_atomics cl_khr_fp64 cl_khr_byte_addressable_store cl_khr_icd cl_khr_gl_sharing cl_nv_compiler_options cl_nv_device_attribute_query cl_nv_pragma_unroll cl_nv_copy_opts cl_nv_create_buffer cl_khr_int64_base_atomics cl_khr_int64_extended_atomics
>   Platform Extensions function suffix             NV
>
>   Platform Name                                   NVIDIA CUDA
> Number of devices                                 1
>   Device Name                                     GeForce 940MX
>   Device Vendor                                   NVIDIA Corporation
>   Device Vendor ID                                0x10de
>   Device Version                                  OpenCL 1.2 CUDA
>   Driver Version                                  440.100
>   Device OpenCL C Version                         OpenCL C 1.2
>   Device Type                                     GPU
>   Device Topology (NV)                            PCI-E, 02:00.0
>   Device Profile                                  FULL_PROFILE
>   Device Available                                Yes
>   Compiler Available                              Yes
>   Linker Available                                Yes
>   Max compute units                               3
>   Max clock frequency                             1241MHz
>   Compute Capability (NV)                         5.0
>   Device Partition                                (core)
>     Max number of sub-devices                     1
>     Supported partition types                     None
>   Max work item dimensions                        3
>   Max work item sizes                             1024x1024x64
>   Max work group size                             1024
>   Preferred work group size multiple              32
>   Warp size (NV)                                  32
>   Preferred / native vector sizes
>     char                                                 1 / 1
>     short                                                1 / 1
>     int                                                  1 / 1
>     long                                                 1 / 1
>     half                                                 0 / 0        (n/a)
>     float                                                1 / 1
>     double                                               1 / 1        (cl_khr_fp64)
>   Half-precision Floating-point support           (n/a)
>   Single-precision Floating-point support         (core)
>     Denormals                                     Yes
>     Infinity and NANs                             Yes
>     Round to nearest                              Yes
>     Round to zero                                 Yes
>     Round to infinity                             Yes
>     IEEE754-2008 fused multiply-add               Yes
>     Support is emulated in software               No
>     Correctly-rounded divide and sqrt operations  Yes
>   Double-precision Floating-point support         (cl_khr_fp64)
>     Denormals                                     Yes
>     Infinity and NANs                             Yes
>     Round to nearest                              Yes
>     Round to zero                                 Yes
>     Round to infinity                             Yes
>     IEEE754-2008 fused multiply-add               Yes
>     Support is emulated in software               No
>   Address bits                                    64, Little-Endian
>   Global memory size                              2101870592 (1.958GiB)
>   Error Correction support                        No
>   Max memory allocation                           525467648 (501.1MiB)
>   Unified memory for Host and Device              No
>   Integrated memory (NV)                          No
>   Minimum alignment for any data type             128 bytes
>   Alignment of base address                       4096 bits (512 bytes)
>   Global Memory cache type                        Read/Write
>   Global Memory cache size                        73728 (72KiB)
>   Global Memory cache line size                   128 bytes
>   Image support                                   Yes
>     Max number of samplers per kernel             32
>     Max size for 1D images from buffer            134217728 pixels
>     Max 1D or 2D image array size                 2048 images
>     Max 2D image size                             16384x16384 pixels
>     Max 3D image size                             4096x4096x4096 pixels
>     Max number of read image args                 256
>     Max number of write image args                16
>   Local memory type                               Local
>   Local memory size                               49152 (48KiB)
>   Registers per block (NV)                        65536
>   Max number of constant args                     9
>   Max constant buffer size                        65536 (64KiB)
>   Max size of kernel argument                     4352 (4.25KiB)
>   Queue properties
>     Out-of-order execution                        Yes
>     Profiling                                     Yes
>   Prefer user sync for interop                    No
>   Profiling timer resolution                      1000ns
>   Execution capabilities
>     Run OpenCL kernels                            Yes
>     Run native kernels                            No
>     Kernel execution timeout (NV)                 Yes
>   Concurrent copy and kernel execution (NV)       Yes
>     Number of async copy engines                  1
>   printf() buffer size                            1048576 (1024KiB)
>   Built-in kernels
>   Device Extensions                               cl_khr_global_int32_base_atomics cl_khr_global_int32_extended_atomics cl_khr_local_int32_base_atomics cl_khr_local_int32_extended_atomics cl_khr_fp64 cl_khr_byte_addressable_store cl_khr_icd cl_khr_gl_sharing cl_nv_compiler_options cl_nv_device_attribute_query cl_nv_pragma_unroll cl_nv_copy_opts cl_nv_create_buffer cl_khr_int64_base_atomics cl_khr_int64_extended_atomics
>
> NULL platform behavior
>   clGetPlatformInfo(NULL, CL_PLATFORM_NAME, ...)  NVIDIA CUDA
>   clGetDeviceIDs(NULL, CL_DEVICE_TYPE_ALL, ...)   Success [NV]
>   clCreateContext(NULL, ...) [default]            Success [NV]
>   clCreateContextFromType(NULL, CL_DEVICE_TYPE_DEFAULT)  No platform
>   clCreateContextFromType(NULL, CL_DEVICE_TYPE_CPU)  No devices found in platform
>   clCreateContextFromType(NULL, CL_DEVICE_TYPE_GPU)  No platform
>   clCreateContextFromType(NULL, CL_DEVICE_TYPE_ACCELERATOR)  No devices found in platform
>   clCreateContextFromType(NULL, CL_DEVICE_TYPE_CUSTOM)  Invalid device type for platform
>   clCreateContextFromType(NULL, CL_DEVICE_TYPE_ALL)  No platform
>
> ICD loader properties
>   ICD loader Name                                 OpenCL ICD Loader
>   ICD loader Vendor                               OCL Icd free software
>   ICD loader Version                              2.2.11
>   ICD loader Profile                              OpenCL 2.1
>

openjournals/joss-reviews#2276
