### 2020-07-05 CentreForHydrology/WDPM#24 @awickert
Update link to WDPM webpage and consider long-term access to documentation
The link included in the `README.md` should be updated to:
https://research-groups.usask.ca/hydrology/modelling/wdpm.php

This also illustrates one difficulty of hosting documentation off-site. I may suggest that you include the manual as part of this repo (perhaps via git LFS, if you are concerned about space + commits). This can then become part of a version of record, published permanently via Zenodo -- or (potentially, I do not know) directly via JOSS.

openjournals/joss-reviews#2276
