### 2020-07-05 CentreForHydrology/WDPM#24 @awickert
This is a partial duplicate of #7.
