### 2020-07-05 CentreForHydrology/WDPM#25 @awickert
Missing CL subdirectory
Overlaps with #16, #20, and/or #23.

Following the compilation instructions given in the manual [here](https://research-groups.usask.ca/hydrology/modelling/wdpm.php), I find that the program is missing a `CL` subdirectory.

```
$ gcc WDPMCL-Command-line.c -lOpenCL -o WDPMCL
WDPMCL-Command-line.c:25:10: fatal error: CL/cl.h: No such file or directory
   25 | #include <CL/cl.h>
      |          ^~~~~~~~~
compilation terminated.
```

openjournals/joss-reviews#2276
