### 2020-07-05 #2276 @awickert
@kbarnhart - I didn't accept the invite in time and it expired, so I cannot update the checklist. Can you re-send? (With apologies!)

I finished a first pass through, and will be able to add more material once I am able to compile the source code.
