### 2020-07-05 CentreForHydrology/WDPM#26 @awickert
Rainfall/snowmelt relative importance
In the paper, you claim:

"Addition of water could be caused by rainfall, but in the Prairies it is more commonly due to snowmelt runoff."

Can you cite a reference showing that this is true? Considering that the formerly-glaciated prairies cover a large north--south extent, I am not sure that this is the case everywhere, or even in most cases.

https://github.com/openjournals/joss-reviews/issues/2276
