### 2020-07-05 CentreForHydrology/WDPM#27 @awickert
Evaporation > infiltration?
You state:

"Subtraction of water may be due to infiltration to the soils, but in the Prairies it is more commonly caused by evaporation."

Could you cite a source demonstrating that this is the case?

https://github.com/openjournals/joss-reviews/issues/2276
