### 2020-07-05 CentreForHydrology/WDPM#28 @awickert
Capitalization of Prairie
Could you explain what makes Prairie a proper noun? Is there a specific prairie region that you have in mind, and/or is this simply convention in Canadian English?

https://github.com/openjournals/joss-reviews/issues/2276
