### 2020-07-13 #2276 @kbarnhart
@KevinShook and @r-barnes @awickert @KCallaghan

I wanted to write a quick message to check in and provide some guidance regarding the JOSS process. As best as I can tell the reviewers have been quite thorough and have opened many issues in the source repository as part of their reviews.

At this point, the next step is for the authors to address these issues and close them when completed. If the authors have clarifying questions for the reviewers or for me (e.g., does this pull request address your concern), please tag us into the conversation. If the authors think that an issue that was opened is out of scope, the issue itself is the place to discuss. As always, please consider me a resource throughout this process.

Once the authors feel like they have addressed the reviewers issues such that the reviewers are able to complete the review checklist, they should comment here indicating this.

Know that I'll check into this thread and the main repository to keep track of progress/status changes at a frequency of about every 1-2 weeks.

Questions, comments, or concerns always welcome.
