### 2020-07-13 #2276 @KevinShook
Thanks very much, this is very helpful for understanding the process.
We'll work on addressing the issues raised.
