### 2020-08-03 #2276 @kbarnhart
@KevinShook I wanted to check in and see how things were going with addressing the issues raised by reviewers. Know that JOSS doesn't have a required timeline for revisions, but it would be helpful for me to know an approximate time frame.
