### 2020-08-03 #2276 @KevinShook
Thanks very much for keeping in touch.
We have almost completed the conversion of the GUI from Python 2 to 3.
Once the tested is complete then the rest of the issues should be fairly simple to address.
I have started editing the paper itself, and the requested clarifications seem to be relatively easily addressed.
Barring unforeseen issues, I would anticipate completion sometime this month.
