### 2020-09-06 679e5ce
Removing accidental debug statement
