### 2020-09-06 c4e9cb1
Support for systems with only one platform
