+++ src/WDPM.py
@@ -1,4 +1,3 @@
-import wx
 import os
 import sys
 import string
@@ -7,20 +6,17 @@ import subprocess
 import threading
 import time
 import textwrap
-import lzma
+
+import wx
+from queue import Queue, Empty  # python 3.x
 
 lock = threading.Lock()
-try:
-    from Queue import Queue, Empty
-except ImportError:
-    from queue import Queue, Empty  # python 3.x
 
-
-if os.path.isfile('runoff.cl')==True:
+if os.path.isfile('runoff.cl'):
     pass
 else:
     raise AssertionError("runoff.cl not present. Exiting now")
-if os.path.isfile('colormap_black.txt')==True:
+if os.path.isfile('colormap_black.txt'):
     pass
 else:
     raise AssertionError("colormap_black.txt not present. Exiting now")
@@ -29,26 +25,15 @@ platx = platform.system()
 
 
 if platx=='Darwin' or platx=='Linux':
-	if os.path.isfile('cmap_black.sh')==True:
+	if os.path.isfile('cmap_black.sh'):
 	    pass
 	else:
 	    raise AssertionError("cmap_black.sh not present. Exiting now")
 
-	if os.path.isfile('WDPMCL')==True:
+	if os.path.isfile('WDPMCL'):
 	    pass
 	else:
 	    raise AssertionError("WDPMCL not present. Exiting now")
-else:
-	import win32com.client
-	if os.path.isfile('cmap.bat')==True:
-	    pass
-	else:
-	    raise AssertionError("cmap.bat not present. Exiting now")
-	
-	if os.path.isfile('WDPMCL.exe')==True:
-	    pass
-	else:
-	    raise AssertionError("WDPMCL.exe not present. Exiting now")
 
 class RedirectText:
 	def __init__(self,aWxTextCtrl):
@@ -59,7 +44,7 @@ class RedirectText:
 
 class CharValidator(wx.PyValidator):
 	def __init__(self,flag):
-		wx.PyValidator.__init__(self)
+		wx.Validator.__init__(self)
 		self.flag=flag
 		self.Bind(wx.EVT_CHAR, self.OnChar)
 
@@ -77,7 +62,7 @@ class CharValidator(wx.PyValidator):
 
 	def OnChar(self, evt):
 		key=chr(evt.GetKeyCode())
-		if self.flag == "no-alpha" and key in string.letters:
+		if self.flag == "no-alpha" and key in string.ascii_letters:
 		    return
 		if self.flag == "no-digit" and key in string.digits:
 		    return
@@ -85,26 +70,26 @@ class CharValidator(wx.PyValidator):
 
 class Size(wx.Frame):
 	def __init__(self, parent, id_):
-		wx.Frame.__init__(self, parent, id_, 'WDPM Program', size=(1366, 768),
+		wx.Frame.__init__(self, parent, id_, 'WDPM Program', size=(1300, 850),
 				  style=wx.MINIMIZE_BOX|wx.MAXIMIZE_BOX|wx.RESIZE_BORDER|wx.SYSTEM_MENU|
 				  wx.CAPTION|wx.CLOSE_BOX|wx.CLIP_CHILDREN)
-		self.panel = wx.PyScrolledWindow(self, id_)
+		self.panel = wx.ScrolledWindow(self, id_)
 		#self.panel=wx.Panel(self)
-		x=25
+		x=35
 		y=5
 		
 		self.lblname9xa = wx.StaticText(self.panel, label="Output to .PNG", pos=(10,23*x))
-		self.txt9xa = wx.TextCtrl(self.panel, -1, pos=(8*x, 23*x), size=(2.5*x, x-y))
[…]
