### 2020-09-08 #2276 @KevinShook
Thanks for getting in touch.
I just merged a pull request today which should have all of the fixes to date.
The GUI has been converted from Python 2 to 3, and no-longer crashes.
I believe that the style issues have also been addressed.
The main outstanding issues are a) there's still one warning left to fix, and b) although we have a makefile, it's not included as it needs a bit of work.
The paper has been updated, but still needs to be gone over, and the manual is being updated as well.
