### 2020-09-13 #2276 @r-barnes
@KevinShook: I'd recommend `cmake` as a better option than a makefile ([example here](https://github.com/r-barnes/Barnes2020-FillSpillMerge/blob/master/CMakeLists.txt)). If you'd like help getting one setup, I'm happy to help.
