### 2020-09-13 #2276 @KevinShook
Thanks very much. We are planning on going with cmake as was recommended by one of the developers.
