### 2020-09-30 #2276 @KevinShook
Thanks for checking in.
We are still finishing a few issues.
We are replacing the makefile with cmake as was suggested by r-barnes (thanks!).
We are also working on having the program work with a single executable. Previously, the GUI called a separate executable, from the command line version, which is obviously not desireable.
One these are done, the manual will need to be updated.
I'm going away for 3 weeks, so will probably not be able to work on it much in the interim.
