### 2020-11-02 #2276 @KevinShook
Thanks. We have finally solved a problem which was responsible for there being 2 different executables. There is now a single executable, but we are working on making it more efficient. When this last hurdle is cleared, we can finally cross off all of the technical issues, and finish the documentation.
