### 2020-11-18 cc878af
Zero threshold depth implemented
