+++ paper.bib
@@ -1,106 +1,0 @@
-
-@incollection{armstrongUSINGWETLANDPONDING2013a,
-  title = {{{USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTIC TOOL FOR PRAIRIE FLOOD HAZARD ASSESSMENT}}},
-  booktitle = {Putting {{Prediction}} in {{Ungauged Basins}} into {{Practice}}},
-  author = {Armstrong, Robert and Kayter, Cameron and Shook, Kevin and Hill, Harvey},
-  year = {2013},
-  pages = {255--270},
-  file = {/home/kevin/data/zotero/storage/NTA77NKG/Armstrong et al. - USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTI.pdf},
-  language = {en}
-}
-
-@article{pomeroyColdRegionsHydrological2007,
-  title = {The Cold Regions Hydrological Model : A Platform for Basing Process Representation and Model Structure on Physical Evidence},
-  author = {Pomeroy, J W and Gray, D M and Brown, T and Hedstrom, N R and Quinton, W L and Granger, R J and Carey, S K},
-  year = {2007},
-  month = sep,
-  volume = {2667},
-  pages = {2650--2667},
-  issn = {08856087},
-  doi = {10.1002/hyp.6787},
-  file = {/home/kevin/data/zotero/storage/PCHIC3ZJ/Pomeroy et al. - 2007 - The cold regions hydrological model a platform for basing process representation and model structure on physica.pdf},
-  journal = {Hydrological Processes},
-  keywords = {accepted 12 december 2006,blowing snow,distributed modelling,evaporation,frozen soil infiltration,hillslope,hydrological modelling,hydrological response unit,physically based modelling,received 1 june 2006,runoff,snowmelt,subsurface flow},
-  number = {19}
-}
-
-@techreport{shapiroMAPCALCAlgebraGIS1992,
-  title = {R.{{MAPCALC}}. {{An Algebra}} for {{GIS}} and {{Image Processing}}},
-  author = {Shapiro, Michael and Westervelt, Jim},
-  year = {1992},
-  pages = {1--22},
-  institution = {{U.S. Army Construction Engineering Researc h Laboratory, Champaign, Illinois 61801, U.S.A.}},
-  file = {/home/kevin/data/zotero/storage/9RD6YKJE/Shapiro, Westervelt - 1992 - An Algebra for GIS and Image Processing.pdf}
-}
-
-@phdthesis{shawInfluenceContributingArea2009,
-  title = {The Influence of Contributing Area on the Hydrology of the Prairie Pothole Region of {{North America}}},
-  author = {Shaw, D A},
-  year = {2009},
-  file = {/home/kevin/data/zotero/storage/9XLPZSQL/Shaw - 2009 - The influence of contributing area on the hydrolog.pdf},
-  type = {{{PhD Thesis}}}
-}
-
-@article{shookMemoryEffectsDepressional2011,
-  title = {Memory Effects of Depressional Storage in {{Northern Prairie}} Hydrology},
-  author = {Shook, Kevin R and Pomeroy, John W},
-  year = {2011},
-  month = dec,
-  volume = {25},
-  pages = {3890--3898},
-  issn = {08856087},
-  doi = {10.1002/hyp.8381},
-  copyright = {All rights reserved},
-  file = {/home/kevin/data/zotero/storage/T4IW4VGJ/Shook and Pomeroy - 2011 - Memory effects of depressional storage in Northern.pdf},
-  journal = {Hydrological Processes},
-  keywords = {accepted 5 october 2011,connectivity,contributing area,hysteresis,memory,prairie,received 20 april 2011,wetland},
[…]
