+++ paper/paper.bib
@@ -1,0 +1,231 @@
+
+@incollection{armstrongUSINGWETLANDPONDING2013a,
+  title = {{{USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTIC TOOL FOR PRAIRIE FLOOD HAZARD ASSESSMENT}}},
+  booktitle = {Putting {{Prediction}} in {{Ungauged Basins}} into {{Practice}}},
+  author = {Armstrong, Robert and Kayter, Cameron and Shook, Kevin and Hill, Harvey},
+  year = {2013},
+  pages = {255--270},
+  file = {/home/kevin/data/zotero/storage/NTA77NKG/Armstrong et al. - USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTI.pdf},
+  language = {en}
+}
+
+@article{elboshyFrameworkPluvialFlood2019,
+  title = {A Framework for Pluvial Flood Risk Assessment in {{Alexandria}} Considering the Coping Capacity},
+  author = {Elboshy, Bahaa and Kanae, Shinjiro and Gamaleldin, Mona and Ayad, Hany and Osaragi, Toshihiro and Elbarki, Waleed},
+  year = {2019},
+  month = mar,
+  volume = {39},
+  pages = {77--94},
+  issn = {2194-5411},
+  doi = {10.1007/s10669-018-9684-7},
+  abstract = {Urbanization and climate change are likely to aggravate the flood risk especially in the developing regions where these are also lack of resources. Risk assessment at the local scale can be seen as an important tool to assist the decision makers to identify and prioritize development, preparedness, and emergency. This paper introduces an integrated framework to assess urban pluvial flood risk, taking into consideration the available coping capacity arrangements as the coping capacity is considered to be the main factor to control the risk impact. The presented framework incorporates the pluvial flood inundation model; the building and social vulnerabilities indices; and coping capacity indicators to identify the risk level in the urban areas and to test the different scenarios for the disaster risk reduction measures. The proposed risk assessment framework has been applied to the city of Alexandria, located in northern Egypt, as there is an increase in pluvial floods in the city causing economic and human losses. A risk map for Almontaza district has been prepared to reveal the risk level for each block, this map can be used for the planning purposes. The introduced framework can increase the efficiency of the preparedness and emergency plans; it can also help the planners to direct the available development resources to the priority areas.},
+  file = {/home/kevin/data/zotero/storage/473FWJR5/Elboshy et al. - 2019 - A framework for pluvial flood risk assessment in A.pdf},
+  journal = {Environ Syst Decis},
+  language = {en},
+  number = {1}
+}
+
+@phdthesis{kissPredictiveMappingWetland2018,
+  title = {Predictive Mapping of Wetland Types and Associated Soils through Digital Elevation Model Analyses in the {{Canadian Prairie Pothole Region}}},
+  author = {Kiss, Jeremy},
+  year = {2018},
+  month = sep,
[…]
