### 2020-11-30 1ff0625
Added CmakeLists.txt, updated manual for building
