+++ paper/paper.md
@@ -16,6 +16,10 @@ authors:
     affiliation: 2
   - name: John W Pomeroy
     affiliation: 1
+	orcid: 0000-0002-4782-7457
+  - name: Tonghe Liu
+    orcid: 0000-0002-2714-3103
+    affiliation: 2
 affiliations:
  - name: Centre for Hydrology, University of Saskatchewan
    index: 1
@@ -27,20 +31,20 @@ bibliography: paper.bib
 
 # Summary
 
-The hydrography of the Prairies of western North America, including Canada, is unusual in that the landscape is flat and recently formed due to the effects of pleistocene glaciation and a semi-arid climate since holocene deglaciation. Therefore, there has not been sufficient energy, time, or runoff water to carve typical dendritic surface water drainage networks in many locations. In these regions, runoff is often detended and sometimes stored by the millions of depressions (known locally as "potholes" or "sloughs") that cover the landscape. Conventional hydrological models are unable to simulate the spatial distribution of ponded water in Prairie basins dominated by depressional storage. When the depressions are filled, the detended water may overflow to another depression, though a process known as "fill and spill" [@spenceHydrologySubarcticCanadian2003]. Therefore, the fraction of a depression-dominated Prairie basin that contributes flow to the outlet will change dynamically with the state of water storage within the basin. This situation is difficult to simulate with conventional hydrological models. 
+The hydrography of the prairies of western North America, including Canada, is unusual in that the landscape is flat and recently formed due to the effects of pleistocene glaciation and a semi-arid climate since holocene deglaciation. Therefore, there has not been sufficient energy, time, or runoff water to carve typical dendritic surface water drainage networks in many locations. In these regions, runoff is often detented and sometimes stored by the millions of depressions (known locally as "potholes" or "sloughs") that cover the landscape. Conventional hydrological models are unable to simulate the spatial distribution of ponded water in prairie basins dominated by depressional storage. When the depressions are filled, the detended water may overflow to another depression, though a process known as "fill and spill" [@spenceHydrologySubarcticCanadian2003]. Therefore, the fraction of a depression-dominated prairie basin that contributes flow to the outlet will change dynamically with the state of water storage within the basin. This situation is difficult to simulate with conventional hydrological models. 
 
 # Program description
 
 The WDPM was developed to solve the problems described above. The model distributes simulated water over a digital elevation model (DEM), which is an array of land surface elevations. The program has three modules:
-- **add**, which adds specified depth of water to the DEM. Addition of water could be caused by rainfall, but in the Prairies it is more commonly due to snowmelt runoff.
-- **subtract**, which subtracts a specified depth of water from the DEM. Subtraction of water may be due to infiltration to the soils, but in the Prairies it is more commonly caused by evaporation.
+- **add**, which adds specified depth of water to the DEM. Addition of water can be caused by rainfall, but in the Canadian Prairies it is dominantly due to the melt of snow trapped in depressions and from runoff from adjacent areas [@hayashiWaterSoluteTransfer1998].
+- **subtract**, which subtracts a specified depth of water from the DEM. Subtraction of water may be caused by infiltration to the soils, but in the Canadian Prairies it is dominantly due evaporation [@hayashiWaterSoluteTransfer1998].
 - **drain**, which drains water from the lowest point in the DEM.
 
[…]
