+++ paper/paper.bib
@@ -9,6 +9,53 @@
   language = {en}
 }
 
+@techreport{assiniboinewatershedstewardshipassociationawsaLandInfrastructureResiliency2012,
+  title = {Land \& {{Infrastructure Resiliency Assessment}} ({{LIRA}}) {{Project Assiniboine River Watershed}}, {{SK}}  - {{Pilot Study FINAL REPORT}}},
+  author = {{Assiniboine Watershed Stewardship Association (AWSA)}},
+  year = {2012},
+  month = sep,
+  pages = {238},
+  institution = {{Assiniboine Watershed Stewardship Association (AWSA)}},
+  url = {http://docs.google.com/viewer?a=v\&pid=sites\&srcid=YXNzaW5pYm9pbmV3YXRlcnNoZWQuY29tfGFzc2luaWJvaW5lLXdhdGVyc2hlZC1zdGV3YXJkc2hpcC1hc3NvY2lhdGlvbnxneDo3N2RiNTVkNTczZTFjNjcy},
+  urldate = {2020-08-04},
+  file = {/home/kevin/data/zotero/storage/5IY2WKA5/LIRA Assiniboine River Watershed FINAL REPORT .pdf}
+}
+
+@article{costaPredictingVariableContributing,
+  title = {Predicting Variable Contributing Areas, Hydrological Connectivity, and Solute Transport Pathways for a {{Canadian Prairie}} Basin},
+  author = {Costa, Diogo and Shook, Kevin and Spence, Chris and Elliott, Jane and Baulch, Helen and Wilson, Henry and Pomeroy, John},
+  volume = {n/a},
+  pages = {e2020WR027984},
+  issn = {1944-7973},
+  doi = {10.1029/2020WR027984},
+  url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2020WR027984},
+  urldate = {2020-11-12},
+  abstract = {In cold agricultural regions, seasonal snowmelt over frozen soils provides the primary source of runoff and transports large nutrient loads downstream. The post-glacial landscape of the Canadian Prairies and Northern Plains of the United States creates challenges for hydrological and water quality modelling. Here, the application of conventional hydrological models is problematic because of cold regions hydrological and chemical processes, the lack of fluvially eroded drainage systems, large non-contributing areas to streamflow and level topography. A new hydrodynamic model was developed to diagnose overland flow from snowmelt in this situation. The model was used to calculate the effect of variable contributing areas on (1) hydrological connectivity and the development of (2) tipping-points in streamflow generation and (3) predominant chemical transport pathways. The agricultural Steppler Basin in Manitoba, Canada, was used to evaluate the model and diagnose snowmelt runoff. Relationships were established between contributing area and (1) snowmelt runoff intensity, (2) seasonal snowmelt volumes and duration, as well as (3) inundated, active and connected areas. Variations in the contributing area depended on terrain and snowmelt characteristics including wind redistribution of snow. Predictors of hydrological response and the size of the contributing area were developed which can be used in larger scale hydrological models of similar regions.},
+  annotation = {\_eprint: https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2020WR027984},
+  copyright = {This article is protected by copyright. All rights reserved.},
+  file = {/home/kevin/data/zotero/storage/SM5KF46T/2020WR027984.html},
+  journal = {Water Resources Research},
+  keywords = {Flooding,hydrodynamic modelling,Prairies,Snowmelt,Transport pathways,Variable contributing areas},
+  language = {en},
+  note = {e2020WR027984 2020WR027984},
+  number = {n/a}
+}
+
[…]
