### 2020-12-07 9de90a6
minor changes to CMakeLists.txt
