+++ src/CMakeLists.txt
@@ -8,27 +8,6 @@ IF (UNIX)
 	target_link_libraries(WDPMCL OpenCL)	
 	find_package(PythonInterp 3 REQUIRED)
 	find_package(PythonLibs 3 REQUIRED)
-
-ELSEIF(WIN32)	
-	MESSAGE(STATUS "curent platform : windows")
-	project(WDPMCL)
-	set(CMAKE_C_COMPILER "gcc")
-	find_package(Python 3)
-	IF(PYTHON_FOUND)
-		message(STATUS "python3 found")
-	ELSE()
-		message(STATUS "python3 not found, please install the python3")
-	ENDIF()
-	ADD_EXECUTABLE(WDPMCL WDPMCL.c)
-	message(STATUS "########################################################################")
-	message(STATUS "please make sure you have set the correct OpenCL directory in cmake file")
-	message(STATUS "########################################################################")
-	#please set the directory of OpenCL library 
-	target_link_libraries(WDPMCL PUBLIC "C:/OpenCL/OpenCL_SDK/3.0/lib/x86/OpenCL.lib")
-	#please set the directory of OpenCL header file
-	target_include_directories(WDPMCL PUBLIC "C:/OpenCL/OpenCL_SDK/3.0/include")
-	set(CMAKE_C_FLAGS "-std=c11")
-	
 ELSEIF(APPLE)
 	MESSAGE(STATUS "current platform: MacOS")
 	project(WDPMCL)
@@ -37,4 +16,23 @@ ELSEIF(APPLE)
 	target_link_libraries(WDPMCL "-framework OpenCL")
 	find_package(PythonInterp 3 REQUIRED)
 	find_package(PythonLibs 3 REQUIRED)
+ELSE()
+	MESSAGE(STATUS "curent platform : windows")
+	project(WDPMCL)
+	set(CMAKE_C_COMPILER "gcc")
+	find_package(Python 3)
+	IF(PYTHON_FOUND)
+		message(STATUS "python3 found")
+	ELSE()
+		message(STATUS "python3 not found, please install python3")
+	ENDIF()
+	ADD_EXECUTABLE(WDPMCL WDPMCL-threshold6.c)
+	message(STATUS "########################################################################")
+	message(STATUS "please make sure you have set the correct OpenCL directory in CMakeLists.txt")
+	message(STATUS "########################################################################")
+	#please set the directory of OpenCL library 
+	target_link_libraries(WDPMCL PUBLIC "C:/OpenCL/OpenCL_SDK/3.0/lib/x86/OpenCL.lib")
+	#please set the directory of OpenCL header file
+	target_include_directories(WDPMCL PUBLIC "C:/OpenCL/OpenCL_SDK/3.0/include")
+	set(CMAKE_C_FLAGS "-std=c11")
 ENDIF()
