### 2021-01-31 CentreForHydrology/WDPM#1 @KevinShook
The GUI has been updated to Python 3 and now appears to work properly
