### 2021-01-31 CentreForHydrology/WDPM#2 @KevinShook
The GUI has been updated to use Python 3 and now works properly. The program can be run with or without the GUI as explained in the manual.
