### 2021-01-31 CentreForHydrology/WDPM#4 @KevinShook
We are getting rid of all binaries other than Windows as we would prefer the users to use the source code.
