### 2021-01-31 CentreForHydrology/WDPM#15 @KevinShook
We have removed all the binary blobs.
