### 2021-01-31 CentreForHydrology/WDPM#16 @KevinShook
We are now using CMake
