### 2021-02-01 CentreForHydrology/WDPM#7 @KevinShook
The documentation has been updated to reflect the new location of other supporting files.
We have removed all binaries from the other website, except for a Windows binary, as we don't expect that
most Windows users will be able to compile the program.
