+++ src/WDPM.py
@@ -1,14 +1,14 @@
 import os
-import sys
-import string
 import platform
+import queue
+import string
 import subprocess
+import sys
+import textwrap
 import threading
 import time
-import textwrap
 
 import wx
-from queue import Queue, Empty  # python 3.x
 
 lock = threading.Lock()
 
@@ -747,10 +747,15 @@ class Size(wx.Frame):
 			else:
 			    print ("DEM file not present. Use the Browse button to locate file.")
 			    plat="error"
-			   
-			cmd = [solver, method, demfilename, waterfilename, wateroutputfilename, 
-				checkpointfilename, waterdeptha, runoffrac, elevationtol, method1, method2, threshold, limitation]
-			self.Module2(cmd)
+			
+			if plat=='Darwin' or plat=='Linux':   
+				cmd = [solver, method, demfilename, waterfilename, wateroutputfilename, 
+					checkpointfilename, waterdeptha, runoffrac, elevationtol, method1, method2, threshold, limitation]
+				self.Module2(cmd)
+			else:
+				cmd = [solverw, method, demfilename, waterfilename, wateroutputfilename, 
+					checkpointfilename, waterdeptha, runoffrac, elevationtol, method1, method2, threshold, limitation]
+				self.Module2(cmd)
 		elif method=='subtract':
 			demfilename = str(self.txt1.GetValue())
 			waterfilename = str(self.txt2.GetValue())
@@ -808,9 +813,14 @@ class Size(wx.Frame):
 			if waterdepths=='':
 			   self.OnErrorDepthS()
 			   plat='error'
-			cmd = [solver, method, demfilename, waterfilename, wateroutputfilename, 
-				checkpointfilename, waterdepths, elevationtol, method1, method2,threshold, limitation]
-			self.Module2(cmd)
+			if plat=='Darwin' or plat=='Linux':
+				cmd = [solver, method, demfilename, waterfilename, wateroutputfilename, 
+					checkpointfilename, waterdepths, elevationtol, method1, method2,threshold, limitation]
+				self.Module2(cmd)
+			else:
+				cmd = [solverw, method, demfilename, waterfilename, wateroutputfilename, 
+					checkpointfilename, waterdepths, elevationtol, method1, method2,threshold, limitation]
+				self.Module2(cmd)
 		elif method=='drain':
 			demfilename = str(self.txt1.GetValue())
 			waterfilename = str(self.txt2.GetValue())
@@ -867,46 +877,25 @@ class Size(wx.Frame):
 			if draintol=='':
 			   self.OnErrorDrain()
 			   plat='error'
-			cmd = [solver, method, demfilename, waterfilename, wateroutputfilename, 
-				checkpointfilename, elevationtol, draintol, method1, method2,threshold, limitation]
-			self.Module2(cmd)
+			if plat=='Darwin' or plat=='Linux':
+				cmd = [solver, method, demfilename, waterfilename, wateroutputfilename, 
+					checkpointfilename, elevationtol, draintol, method1, method2,threshold, limitation]
+				self.Module2(cmd)
+			else:
+				cmd = [solverw, method, demfilename, waterfilename, wateroutputfilename, 
+					checkpointfilename, elevationtol, draintol, method1, method2,threshold, limitation]
+				self.Module2(cmd)
 		elif method=='TextFile':
 			filename = str(self.txt9a.GetValue())
 			if filename=='':
 				self.OnErrorFile()
 			else:
-				newfile1 = os.path.join(self.txt0a.GetValue(),"input1.in")
-				newfile2 = os.path.join(self.txt0a.GetValue(),"input2.in")
-				f = open(filename,'r')
-				data_list = f.readlines()
[…]
