+++ src/WDPM-GUI.txt
@@ -1,0 +1,975 @@
+import os
+import platform
+import queue
+import string
+import subprocess
+import sys
+import textwrap
+import threading
+import time
+
+import wx
+
+lock = threading.Lock()
+
+if os.path.isfile('runoff.cl'):
+    pass
+else:
+    raise AssertionError("runoff.cl not present. Exiting now")
+if os.path.isfile('colormap_black.txt'):
+    pass
+else:
+    raise AssertionError("colormap_black.txt not present. Exiting now")
+
+platx = platform.system()
+
+
+if platx=='Darwin' or platx=='Linux':
+	if os.path.isfile('cmap_black.sh'):
+	    pass
+	else:
+	    raise AssertionError("cmap_black.sh not present. Exiting now")
+
+	if os.path.isfile('WDPMCL'):
+	    pass
+	else:
+	    raise AssertionError("WDPMCL not present. Exiting now")
+
+class RedirectText:
+	def __init__(self,aWxTextCtrl):
+	    self.out=aWxTextCtrl
+
+	def write(self,string):
+	    wx.CallAfter(self.out.WriteText, string)
+
+class CharValidator(wx.PyValidator):
+	def __init__(self,flag):
+		wx.Validator.__init__(self)
+		self.flag=flag
+		self.Bind(wx.EVT_CHAR, self.OnChar)
+
+	def Clone(self):
+		return CharValidator(self.flag)
+
+	def Validate(self,win):
+		return True
+
+	def TransferToWindow(self):
+		return True
+
+	def TransferFromWindow(self):
+		return True
+
+	def OnChar(self, evt):
+		key=chr(evt.GetKeyCode())
+		if self.flag == "no-alpha" and key in string.ascii_letters:
+		    return
+		if self.flag == "no-digit" and key in string.digits:
+		    return
+		evt.Skip()
+
+class Size(wx.Frame):
+	def __init__(self, parent, id_):
+		wx.Frame.__init__(self, parent, id_, 'WDPM Program', size=(1300, 850),
+				  style=wx.MINIMIZE_BOX|wx.MAXIMIZE_BOX|wx.RESIZE_BORDER|wx.SYSTEM_MENU|
+				  wx.CAPTION|wx.CLOSE_BOX|wx.CLIP_CHILDREN)
+		self.panel = wx.ScrolledWindow(self, id_)
+		#self.panel=wx.Panel(self)
+		x=35
+		y=5
+		
+		self.lblname9xa = wx.StaticText(self.panel, label="Output to .PNG", pos=(10,25*x))
+		self.txt9xa = wx.TextCtrl(self.panel, -1, pos=(round(8*x), round(25*x)), size=(round(2.5*x), round(x-y)))
+		self.button19xa = wx.Button(self.panel, -1, "Browse", pos=(round(10.5*x),round(25*x)), size=(round(2.5*x), round(x-y)))
+		self.button19xa.Bind(wx.EVT_BUTTON, self.OnOpenDEM)
+		self.Convert=wx.Button(self.panel, label="Convert", pos=(round(13*x),round(25*x)), size=(round(2.5*x), round(x-y)))
+		self.Convert.Bind(wx.EVT_BUTTON, self.BitmapConvert)
+		self.txt9xa.Enable(False)
+		self.button19xa.Enable(False)
+		self.Convert.Enable(False)
+		self.runbutton=wx.Button(self.panel, label="Start", pos=(round(10),round(24*x)), size=(round(2.5*x), round(x-y)))
+		self.clearbutton=wx.Button(self.panel, label="Clear", pos=(8*x,24*x), size=(round(2.5*x), x-y))
+		self.endbutton=wx.Button(self.panel, label="End", pos=(13*x,24*x), size=(round(2.5*x), x-y))
[…]
