### 2021-02-03 CentreForHydrology/WDPM#26 @KevinShook
This is certainly the case in the Canadian Prairies where the majority of annual runoff is caused by snowmelt and the depressions actively trap blowing snow. This has been clarified and a supporting reference has been added.
