### 2021-02-03 CentreForHydrology/WDPM#27 @KevinShook
In the Canadian Prairies, depressions are commonly underlain by deep deposits of glacial till, with very low permeabilities.
A supporting reference has been added.
