### 2021-02-03 CentreForHydrology/WDPM#28 @KevinShook
We typically capitalise prairie when we are referring to the Canadian Prairies (a proper noun) rather than prairies in general. This has been tidied to make sure that we are using the appropriate capitalization.
