### 2021-02-03 CentreForHydrology/WDPM#24 @KevinShook
That's a great suggestion. We've added the manual files to the repo.
