### 2021-02-03 CentreForHydrology/WDPM#23 @KevinShook
Getting OpenCL to work is non-trivial. We have removed the binary blobs and now provide some links to
to the appropriate drivers. What OS were you installing for? I have found issues using the Linus Mint OpenCl drivers
in the default repositories and could only get OpenCL to work when I downloaded the Intel drivers directly.
