### 2021-02-03 01d10e3
updated examples in manual, cmap_black.sh now allows spaces in path
