### 2021-02-03 CentreForHydrology/WDPM#23 @KCallaghan
I am also using Linux Mint, and I probably did use the option from the default repository. It sounds like my OpenCL installation is most likely the issue, then.
