### 2021-02-03 CentreForHydrology/WDPM#23 @KevinShook
I had the same issue, as I just did a clean install of Mint 20.1. The Intel drivers in the repository ("beignet") do not work. When I downloaded the drivers from Intel, they would not install unless I did a complete removal of beignet. I suppose the Intel drivers were looking for the beignet configuration files. I'm going to add this information to the manual.
