### 2021-02-03 1de06d4
fixed a few typos, deleted un-needed file
