### 2021-02-03 e61d500
minor typo fixes
