### 2021-02-03 CentreForHydrology/WDPM#20 @KevinShook
No longer getting compilation errors.
Output of make -pedantic does not reveal any of the problems
found before.
