+++ src/requirements.txt
@@ -1,0 +1,1 @@
+wxPython==4.1.0
