### 2021-02-03 88c25b6
added cmap.bat and tidied README.txt
