### 2021-02-03 CentreForHydrology/WDPM#19 @KevinShook
A requirements.txt file has been added, and the external software required is now referenced in README.md as well as in the program manual.
