### 2021-02-04 CentreForHydrology/WDPM#25 @KevinShook
This has now been fixed. The directory will be correctly accessed when the OpenCl drivers are installed.
