### 2021-02-05 #2276 @KevinShook
Thanks for keeping in touch.
I have replied to all but 3 issues, and I believe that all but these last 3 have been solved.
I didn't close the issues, as I assumed that this would be up to the person creating them.

The program now runs (much better thanks to the reviewers), and the GUI has also been greatly
improved.
The last 3 issues (#18, #14, and #13) are concerned with programming style and are being addressed right now.
They should be finished in the next few days.
