### 2021-02-07 #2276 @kbarnhart
@KevinShook Thanks for this update. Glad to know that the review comments have improved the software. I concur with you that the reviewers who created the issues should be the people who close them.

Once you have finished addressing the final issues, please ping me and the reviewers here to let them know that it is time for them to complete their reviews.
