### 2021-02-07 259ac89
Update WDPM.py
1，Fixing the retract issue, and some errors when compiling with pylint
2,  Abstract the UI control (line 341 to line 389), please check
3, For issue 14. Seems that the "repeated myself" part has been fixed. Please let me know if that is not the case.
