+++ src/WDPM.py
@@ -5,20 +5,19 @@ Developed by Oluwaseun Sharomi, Raymond Spiteri and Tonghe Liu
 Numerical Simulation Laboratory, University of Saskatchewan.
 
 --------------------------------------------------------------------
-                                                                   
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 3 of the License, or
 (at your option) any later version.
-                                                                   
+
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
-                                                                   
+
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
-                                                                   
+
 This program adds water to an ArcGIS ASCII file of water runoff
 and redistributes water over the DEM
 From the algorithm of Shapiro, M., & Westervelt, J. (1992).
@@ -54,944 +53,910 @@ platx = platform.system()
 
 
 if platx=='Darwin' or platx=='Linux':
-	if os.path.isfile('cmap_black.sh'):
-	    pass
-	else:
-	    raise AssertionError("cmap_black.sh not present. Exiting now")
+    if os.path.isfile('cmap_black.sh'):
+        pass
+    else:
+        raise AssertionError("cmap_black.sh not present. Exiting now")
 
-	if os.path.isfile('WDPMCL'):
-	    pass
-	else:
-	    raise AssertionError("WDPMCL not present. Exiting now")
+    if os.path.isfile('WDPMCL'):
+        pass
+    else:
+        raise AssertionError("WDPMCL not present. Exiting now")
 
 class RedirectText:
-	def __init__(self,aWxTextCtrl):
-	    self.out=aWxTextCtrl
+    def __init__(self,aWxTextCtrl):
+        self.out=aWxTextCtrl
 
-	def write(self,string):
-	    wx.CallAfter(self.out.WriteText, string)
+    def write(self,string):
+        wx.CallAfter(self.out.WriteText, string)
 
 class CharValidator(wx.PyValidator):
-	def __init__(self,flag):
-		wx.Validator.__init__(self)
-		self.flag=flag
-		self.Bind(wx.EVT_CHAR, self.OnChar)
+    def __init__(self,flag):
+        wx.Validator.__init__(self)
+        self.flag=flag
+        self.Bind(wx.EVT_CHAR, self.OnChar)
 
-	def Clone(self):
-		return CharValidator(self.flag)
+    def Clone(self):
+        return CharValidator(self.flag)
 
-	def Validate(self,win):
-		return True
+    def Validate(self,win):
+        return True
 
-	def TransferToWindow(self):
-		return True
+    def TransferToWindow(self):
+        return True
 
-	def TransferFromWindow(self):
-		return True
+    def TransferFromWindow(self):
+        return True
 
-	def OnChar(self, evt):
[…]
