### 2021-02-08 CentreForHydrology/WDPM#12 @KevinShook
The code has been cleaned up and we have removed the duplications
