+++ src/WDPM.py
@@ -28,11 +28,9 @@ An Algebra for GIS and Image Processing (pp. 1-22).
 
 import os
 import platform
-import queue
 import string
 import subprocess
 import sys
-import textwrap
 import threading
 import time
 
@@ -63,32 +61,25 @@ if platx=='Darwin' or platx=='Linux':
     else:
         raise AssertionError("WDPMCL not present. Exiting now")
 
+
 class RedirectText:
+    '''redirect text'''
     def __init__(self,aWxTextCtrl):
         self.out=aWxTextCtrl
 
     def write(self,string):
+        '''write'''
         wx.CallAfter(self.out.WriteText, string)
 
 class CharValidator(wx.PyValidator):
+    '''check if the char is validate'''
     def __init__(self,flag):
         wx.Validator.__init__(self)
         self.flag=flag
-        self.Bind(wx.EVT_CHAR, self.OnChar)
+        self.Bind(wx.EVT_CHAR, self.on_char)
 
-    def Clone(self):
-        return CharValidator(self.flag)
-
-    def Validate(self,win):
-        return True
-
-    def TransferToWindow(self):
-        return True
-
-    def TransferFromWindow(self):
-        return True
-
-    def OnChar(self, evt):
+    def on_char(self, evt):
+        '''on char'''
         key=chr(evt.GetKeyCode())
         if self.flag == "no-alpha" and key in string.ascii_letters:
             return
@@ -97,6 +88,7 @@ class CharValidator(wx.PyValidator):
         evt.Skip()
 
 class Size(wx.Frame):
+    '''main body'''
     def __init__(self, parent, id_):
         wx.Frame.__init__(self, parent, id_, 'Wetland DEM Ponding Model', size=(1300, 850),
                                   style=wx.MINIMIZE_BOX|wx.MAXIMIZE_BOX|wx.RESIZE_BORDER|
@@ -111,10 +103,10 @@ class Size(wx.Frame):
                         size=(round(2.5*x), round(x-y)))
         self.button19xa = wx.Button(self.panel, -1, "Browse", pos=(round(10.5*x),round(25*x)),
                         size=(round(2.5*x), round(x-y)))
-        self.button19xa.Bind(wx.EVT_BUTTON, self.OnOpenDEM)
+        self.button19xa.Bind(wx.EVT_BUTTON, self.on_open_dem)
         self.Convert=wx.Button(self.panel, label="Convert", pos=(round(13*x),round(25*x)),
                         size=(round(2.5*x), round(x-y)))
-        self.Convert.Bind(wx.EVT_BUTTON, self.BitmapConvert)
+        self.Convert.Bind(wx.EVT_BUTTON, self.bitmap_convert)
         self.txt9xa.Enable(False)
         self.button19xa.Enable(False)
         self.Convert.Enable(False)
@@ -125,12 +117,12 @@ class Size(wx.Frame):
         self.clearbutton=wx.Button(self.panel, label="Clear",
         		pos=(8*x,24*x), size=(round(2.5*x), x-y))
         self.endbutton=wx.Button(self.panel, label="End", pos=(13*x,24*x), size=(round(2.5*x), x-y))
-        self.Bind(wx.EVT_BUTTON, self.RunSim, self.runbutton)
-        self.Bind(wx.EVT_BUTTON, self.OnClearScreen, self.clearbutton)
-        self.Bind(wx.EVT_BUTTON, self.OnAbortButton, self.endbutton)
+        self.Bind(wx.EVT_BUTTON, self.run_sim, self.runbutton)
+        self.Bind(wx.EVT_BUTTON, self.on_clear_screen, self.clearbutton)
[…]
