+++ src/WDPM.py
@@ -96,15 +96,15 @@ class Size(wx.Frame):
         y_p=5
 
         ## Generate picture (text and button)
-        self.lblname9xa = wx.StaticText(self.panel, label="Output to .PNG", pos=(10,25*x))
-        self.txt9xa = wx.TextCtrl(self.panel, -1, pos=(round(8*x), round(25*x)),
-                        size=(round(2.5*x), round(x-y)))
-        self.button19xa = wx.Button(self.panel, -1, "Browse", pos=(round(10.5*x),round(25*x)),
-                        size=(round(2.5*x), round(x-y)))
+        self.lblname9xa = wx.StaticText(self.panel, label="Output to .PNG", pos=(10,25*x_p))
+        self.txt9xa = wx.TextCtrl(self.panel, -1, pos=(round(8*x_p), round(25*x_p)),
+                        size=(round(2.5*x_p), round(x_p-y_p)))
+        self.button19xa = wx.Button(self.panel, -1, "Browse", pos=(round(10.5*x_p),round(25*x_p)),
+                        size=(round(2.5*x_p), round(x_p-y_p)))
         self.button19xa.Bind(wx.EVT_BUTTON, self.on_open_dem)
-        self.Convert=wx.Button(self.panel, label="Convert", pos=(round(13*x),round(25*x)),
-                        size=(round(2.5*x), round(x-y)))
-        self.Convert.Bind(wx.EVT_BUTTON, self.bitmap_convert)
+        self.convert=wx.Button(self.panel, label="convert", pos=(round(13*x_p),round(25*x_p)),
+                        size=(round(2.5*x_p), round(x_p-y_p)))
+        self.convert.Bind(wx.EVT_BUTTON, self.bitmap_convert)
         self.txt9xa.Enable(False)
         self.button19xa.Enable(False)
         self.convert.Enable(False)
@@ -113,8 +113,9 @@ class Size(wx.Frame):
         self.runbutton=wx.Button(self.panel, label="Start", pos=(round(10),round(24*x_p)),
                         size=(round(2.5*x_p), round(x_p-y_p)))
         self.clearbutton=wx.Button(self.panel, label="Clear",
-        		pos=(8*x,24*x), size=(round(2.5*x), x-y))
-        self.endbutton=wx.Button(self.panel, label="End", pos=(13*x,24*x), size=(round(2.5*x), x-y))
+        		pos=(8*x_p,24*x_p), size=(round(2.5*x_p), x_p-y_p))
+        self.endbutton=wx.Button(self.panel, label="End", pos=(13*x_p,24*x_p),
+        		size=(round(2.5*x_p), x_p-y_p))
         self.Bind(wx.EVT_BUTTON, self.run_sim, self.runbutton)
         self.Bind(wx.EVT_BUTTON, self.on_clear_screen, self.clearbutton)
         self.Bind(wx.EVT_BUTTON, self.on_abort_button, self.endbutton)
@@ -137,9 +138,10 @@ class Size(wx.Frame):
         self.Show(True)
 
         ## Working directory (text and button)
-        self.lblname0a = wx.StaticText(self.panel, label="Working Directory:", pos=(10,x))
-        self.txt0a = wx.TextCtrl(self.panel, -1, pos=(8*x, x), size=(5*x, x-y))
-        self.button00a = wx.Button(self.panel, -1, "Browse", pos=(13*x,x), size=(round(2.5*x), x-y))
+        self.lblname0a = wx.StaticText(self.panel, label="Working Directory:", pos=(10,x_p))
[…]
