### 2021-02-09 a8de744
deleted pylint output
