+++ src/pylint.txt
@@ -1,85 +1,0 @@
-************* Module WDPM
-WDPM.py:696:0: W1401: Anomalous backslash in string: '\W'. String constant might be missing an r prefix. (anomalous-backslash-in-string)
-WDPM.py:1:0: C0103: Module name "WDPM" doesn't conform to snake_case naming style (invalid-name)
-WDPM.py:53:3: R1714: Consider merging these comparisons with "in" to "platx in ('Darwin', 'Linux')" (consider-using-in)
-WDPM.py:70:19: W0621: Redefining name 'string' from outer scope (line 31) (redefined-outer-name)
-WDPM.py:65:0: R0903: Too few public methods (1/2) (too-few-public-methods)
-WDPM.py:77:8: W0233: __init__ method from a non direct base class 'Validator' is called (non-parent-init-called)
-WDPM.py:74:0: R0903: Too few public methods (1/2) (too-few-public-methods)
-WDPM.py:107:8: C0103: Attribute name "Convert" doesn't conform to snake_case naming style (invalid-name)
-WDPM.py:90:0: R0901: Too many ancestors (8/7) (too-many-ancestors)
-WDPM.py:90:0: R0902: Too many instance attributes (71/7) (too-many-instance-attributes)
-WDPM.py:131:8: W0622: Redefining built-in 'quit' (redefined-builtin)
-WDPM.py:97:8: C0103: Variable name "x" doesn't conform to snake_case naming style (invalid-name)
-WDPM.py:98:8: C0103: Variable name "y" doesn't conform to snake_case naming style (invalid-name)
-WDPM.py:92:4: R0915: Too many statements (129/50) (too-many-statements)
-WDPM.py:276:23: W0613: Unused argument 'event' (unused-argument)
-WDPM.py:314:4: R0201: Method could be a function (no-self-use)
-WDPM.py:322:4: R0201: Method could be a function (no-self-use)
-WDPM.py:331:22: W0613: Unused argument 'event' (unused-argument)
-WDPM.py:335:4: E0202: An attribute defined in WDPM line 631 hides this method (method-hidden)
-WDPM.py:335:22: W0613: Unused argument 'event' (unused-argument)
-WDPM.py:348:4: R0913: Too many arguments (28/5) (too-many-arguments)
-WDPM.py:348:4: R0914: Too many local variables (28/15) (too-many-locals)
-WDPM.py:381:21: W0613: Unused argument 'event' (unused-argument)
-WDPM.py:405:4: R0201: Method could be a function (no-self-use)
-WDPM.py:414:4: R0201: Method could be a function (no-self-use)
-WDPM.py:425:4: R0201: Method could be a function (no-self-use)
-WDPM.py:434:4: R0201: Method could be a function (no-self-use)
-WDPM.py:446:4: R0201: Method could be a function (no-self-use)
-WDPM.py:455:4: R0201: Method could be a function (no-self-use)
-WDPM.py:464:4: R0201: Method could be a function (no-self-use)
-WDPM.py:473:4: R0201: Method could be a function (no-self-use)
-WDPM.py:482:4: R0201: Method could be a function (no-self-use)
-WDPM.py:491:4: R0201: Method could be a function (no-self-use)
-WDPM.py:500:4: R0201: Method could be a function (no-self-use)
-WDPM.py:509:4: R0201: Method could be a function (no-self-use)
-WDPM.py:518:4: R0201: Method could be a function (no-self-use)
-WDPM.py:564:30: W0613: Unused argument 'event' (unused-argument)
-WDPM.py:573:30: W0613: Unused argument 'event' (unused-argument)
-WDPM.py:590:19: W0703: Catching too general exception Exception (broad-except)
[…]
