### 2021-02-09 CentreForHydrology/WDPM#18 @KevinShook
We have used Pylint to clean up the code, and it is looking much cleaner and easier to follow.
