### 2021-02-12 43b5920
Update CMakeLists.txt
update to work for MacOS
