+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required (VERSION 3.9)
 
-IF (UNIX)
+IF (UNIX AND NOT APPLE)
 	MESSAGE(STATUS "current platform : Linux")
 	project(WDPMCL)
 	SET(CMAKE_C_COMPILER "gcc")
@@ -24,7 +24,7 @@ ELSE()
 	IF(PYTHON_FOUND)
 		message(STATUS "python3 found")
 	ELSE()
-		message(STATUS "python3 not found, please install python3")
+		message(STATUS "python3 not found, please install the python3")
 	ENDIF()
 	ADD_EXECUTABLE(WDPMCL WDPMCL-threshold6.c)
 	message(STATUS "########################################################################")
