### 2021-02-16 0b0f890
edits to documentation
