### 2021-02-16 895d7e0
edits to documentation
