### 2021-02-22 CentreForHydrology/WDPM#25 @awickert
I can confirm that WDPM now successfully compiles (with the help of cmake) on my computer following the instructions in the manual.
