### 2021-02-23 aed9d8e
Update WDPM.py
Delete those (true false message).
