+++ src/WDPM.py
@@ -373,33 +373,33 @@ class Size(wx.Frame):
     		editname7b, lblname9a, txt9a, button19a, txt1, button11, txt2, button12, txt3,
     		txt4, lblname8, combo8, button18, lblname9):
         '''disable or enable the button'''
-        print(self.lblname5.Enable(True) if lblname5==1 else self.lblname5.Enable(False))
-        print(self.editname5.Enable(True) if editname5==1 else self.editname5.Enable(False))
-        print(self.lblname6.Enable(True) if lblname6==1 else self.lblname6.Enable(False))
-        print(self.editname6.Enable(True) if editname6==1 else self.editname6.Enable(False))
-        print(self.lblname7.Enable(True) if lblname7==1 else self.lblname7.Enable(False))
-        print(self.editname7.Enable(True) if editname7==1 else self.editname7.Enable(False))
-        print(self.lblname5a.Enable(True) if lblname5a==1 else self.lblname5a.Enable(False))
-        print(self.editname5a.Enable(True) if editname5a==1 else self.editname5a.Enable(False))
-        print(self.lblname7a.Enable(True) if lblname7a==1 else self.lblname7a.Enable(False))
-        print(self.editname7a.Enable(True) if editname7a==1 else self.editname7a.Enable(False))
-        print(self.lblname6b.Enable(True) if lblname6b==1 else self.lblname6b.Enable(False))
-        print(self.editname6b.Enable(True) if editname6b==1 else self.editname6b.Enable(False))
-        print(self.lblname7b.Enable(True) if lblname7b==1 else self.lblname7b.Enable(False))
-        print(self.editname7b.Enable(True) if editname7b==1 else self.editname7b.Enable(False))
-        print(self.lblname9a.Enable(True) if lblname9a==1 else self.lblname9a.Enable(False))
-        print(self.txt9a.Enable(True) if txt9a==1 else self.txt9a.Enable(False))
-        print(self.button19a.Enable(True) if button19a==1 else self.button19a.Enable(False))
-        print(self.txt1.Enable(True) if txt1==1 else self.txt1.Enable(False))
-        print(self.button11.Enable(True) if button11==1 else self.button11.Enable(False))
-        print(self.txt2.Enable(True) if txt2==1 else self.txt2.Enable(False))
-        print(self.button12.Enable(True) if button12==1 else self.button12.Enable(False))
-        print(self.txt3.Enable(True) if txt3==1 else self.txt3.Enable(False))
-        print(self.txt4.Enable(True) if txt4==1 else self.txt4.Enable(False))
-        print(self.lblname8.Enable(True) if lblname8==1 else self.lblname8.Enable(False))
-        print(self.combo8.Enable(True) if combo8==1 else self.combo8.Enable(False))
-        print(self.button18.Enable(True) if button18==1 else self.button18.Enable(False))
-        print(self.lblname9.Enable(True) if lblname9==1 else self.lblname9.Enable(False))
+        self.lblname5.Enable(True) if lblname5==1 else self.lblname5.Enable(False)
+        self.editname5.Enable(True) if editname5==1 else self.editname5.Enable(False)
+        self.lblname6.Enable(True) if lblname6==1 else self.lblname6.Enable(False)
+        self.editname6.Enable(True) if editname6==1 else self.editname6.Enable(False)
+        self.lblname7.Enable(True) if lblname7==1 else self.lblname7.Enable(False)
+        self.editname7.Enable(True) if editname7==1 else self.editname7.Enable(False)
+        self.lblname5a.Enable(True) if lblname5a==1 else self.lblname5a.Enable(False)
[…]
