### 2021-02-23 #2234 @kthyng
Hi @KevinShook! You can test out your paper compilation, see here: https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles
