+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @incollection{armstrongUSINGWETLANDPONDING2013a,
   title = {{{USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTIC TOOL FOR PRAIRIE FLOOD HAZARD ASSESSMENT}}},
   booktitle = {Putting {{Prediction}} in {{Ungauged Basins}} into {{Practice}}},
-  author = {Armstrong, Robert and Kayter, Cameron and Shook, Kevin and Hill, Harvey},
+  author = {Armstrong, Robert and Kayter, Cameron and Shook, Kevin R and Hill, Harvey},
   year = {2013},
   pages = {255--270},
   file = {/home/kevin/data/zotero/storage/NTA77NKG/Armstrong et al. - USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTI.pdf},
@@ -23,7 +23,7 @@
 
 @article{costaPredictingVariableContributing2020,
   title = {Predicting {{Variable Contributing Areas}}, {{Hydrological Connectivity}}, and {{Solute Transport Pathways}} for a {{Canadian Prairie Basin}}},
-  author = {Costa, Diogo and Shook, Kevin and Spence, Chris and Elliott, Jane and Baulch, Helen and Wilson, Henry and Pomeroy, John W.},
+  author = {Costa, Diogo and Shook, Kevin R and Spence, Chris and Elliott, Jane and Baulch, Helen and Wilson, Henry and Pomeroy, John W.},
   year = {2020},
   volume = {56},
   pages = {e2020WR027984},
@@ -225,7 +225,7 @@
 
 @article{shookStorageDynamicsSimulations2013,
   title = {Storage Dynamics Simulations in Prairie Wetland Hydrology Models: Evaluation and Parameterization},
-  author = {Shook, Kevin and Pomeroy, John W and Spence, Christopher and Boychuk, Lyle},
+  author = {Shook, Kevin R and Pomeroy, John W and Spence, Christopher and Boychuk, Lyle},
   year = {2013},
   month = jun,
   volume = {27},
@@ -242,7 +242,7 @@
 
 @article{shookTransformationFrequencyDistributions2015,
   title = {The Transformation of Frequency Distributions of Winter Precipitation to Spring Streamflow Probabilities in Cold Regions; Case Studies from the {{Canadian Prairies}}},
-  author = {Shook, Kevin and Pomeroy, John and {van der Kamp}, Garth},
+  author = {Shook, Kevin R and Pomeroy, John and {van der Kamp}, Garth},
   year = {2015},
   month = feb,
   volume = {521},
