### 2021-03-26 63deb16
added missing DOIs
