+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @incollection{armstrongUSINGWETLANDPONDING2013a,
   title = {{{USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTIC TOOL FOR PRAIRIE FLOOD HAZARD ASSESSMENT}}},
   booktitle = {Putting {{Prediction}} in {{Ungauged Basins}} into {{Practice}}},
-  author = {Armstrong, Robert and Kayter, Cameron and Shook, Kevin R and Hill, Harvey},
+  author = {Armstrong, Robert and Kayter, Cameron and Shook, Kevin and Hill, Harvey},
   year = {2013},
   pages = {255--270},
   file = {/home/kevin/data/zotero/storage/NTA77NKG/Armstrong et al. - USING THE WETLAND DEM PONDING MODEL AS A DIAGNOSTI.pdf},
@@ -23,7 +23,7 @@
 
 @article{costaPredictingVariableContributing2020,
   title = {Predicting {{Variable Contributing Areas}}, {{Hydrological Connectivity}}, and {{Solute Transport Pathways}} for a {{Canadian Prairie Basin}}},
-  author = {Costa, Diogo and Shook, Kevin R and Spence, Chris and Elliott, Jane and Baulch, Helen and Wilson, Henry and Pomeroy, John W.},
+  author = {Costa, Diogo and Shook, Kevin and Spence, Chris and Elliott, Jane and Baulch, Helen and Wilson, Henry and Pomeroy, John W.},
   year = {2020},
   volume = {56},
   pages = {e2020WR027984},
@@ -80,6 +80,7 @@
   year = {1998},
   volume = {207},
   pages = {42--55},
+  doi = {/10.1016/S0022-1694(98)00098-5},
   file = {/home/kevin/data/zotero/storage/U9UB5RNM/Hayashi, Kamp, Rudolph - 1998 - Water and solute transfer between a prairie wetland and adjacent uplands, 1. Water balance.pdf},
   journal = {Journal of Hydrology}
 }
@@ -225,7 +226,7 @@
 
 @article{shookStorageDynamicsSimulations2013,
   title = {Storage Dynamics Simulations in Prairie Wetland Hydrology Models: Evaluation and Parameterization},
-  author = {Shook, Kevin R and Pomeroy, John W and Spence, Christopher and Boychuk, Lyle},
+  author = {Shook, Kevin and Pomeroy, John W and Spence, Christopher and Boychuk, Lyle},
   year = {2013},
   month = jun,
   volume = {27},
@@ -242,7 +243,7 @@
 
 @article{shookTransformationFrequencyDistributions2015,
   title = {The Transformation of Frequency Distributions of Winter Precipitation to Spring Streamflow Probabilities in Cold Regions; Case Studies from the {{Canadian Prairies}}},
-  author = {Shook, Kevin R and Pomeroy, John and {van der Kamp}, Garth},
+  author = {Shook, Kevin and Pomeroy, John and {van der Kamp}, Garth},
   year = {2015},
   month = feb,
   volume = {521},
@@ -289,6 +290,7 @@
   volume = {11},
   pages = {2486},
   publisher = {{Multidisciplinary Digital Publishing Institute}},
+  doi = {10.3390/w11122486},
   file = {/home/kevin/data/zotero/storage/B7ASH2P7/Thapa et al. - 2019 - “Garbage in, Garbage Out” Does Not Hold True for I.pdf;/home/kevin/data/zotero/storage/MWITRIG6/2486.html},
   journal = {Water},
   number = {12}
@@ -322,6 +324,7 @@
   year = {1993},
   volume = {146},
[…]
