### 2021-03-28 #2276 @KevinShook
A strange (but welcome) thing just happened. Out of the blue, Oluwaseun Sharomi contacted us!
It turns out he had moved from Abu Dhabi to the USA, which was why we were unable to contact him.
He has given his consent to be a co-author, so I have added him to the paper.
