### 2021-03-28 #2276 @awickert
Hi @KevinShook. First off, fantastic news about Oluwaseun Sharomi. This is really an ideal solution. Second, once you sent your message above about completion, I began to tick off the remaining items. One that seemed a sticky point that I should have addressed earlier is "State of the field: Do the authors describe how this software compares to other commonly-used packages?". I have not seen much on this; it might be good to include a paragraph or so. (If I've missed something in my new skim, please just note this.)

Otherwise, I simply need to complete my own tests using it, which I hope to do by the end of next week (sorry for the bit of time, my schedule is packed). So far, so good; hoping that I'll be able to simply tick the boxes, but doing the due diligence.
