### 2021-03-29 #2276 @KevinShook
HI @awickert
We're very relieved to be able to include Oluwaseun Sharomi. As you said, this is by far the best solution.

As for your point, we did describe the differences between WDPM and other types of programs which are used in the region.

We first explained that WDPM is not a hydrological model:

> WDPM does not attempt to compute the magnitudes of the additive or subtractive fluxes, which must be established by measurement or through the use of a physically based model of prairie hydrology such as the Cold Regions Hydrological Modelling platform (CRHM) [@pomeroyColdRegionsHydrological2007].

We then explained the differences between WDPM and hydraulic models, in that WDPM is a model of the state of the
landscape after the fluxes have completed:

> The  model output is the depth of water over each cell of the DEM. WDPM is not a hydraulic model; it is a model of the state of the distribution of water in a landscape after the fluxes have been redistributed. FLUXOS-OVERFLOW is a 2-D hydrodynamic model capable of modelling overland flow rates in the prairies [@costaPredictingVariableContributing2020], but it is much slower than WDPM, and it requires more parameters and more forcing data.

I'm not sure if this is sufficient, or we need to add some further explanations.
