### 2021-04-04 #2276 @awickert
Hi @KevinShook. Regarding your background, you do indeed compare your work to different hydrological models. However, there are a number of computational tools that are quite a lot more similar to what you are proposing here. Some references pertaining to the Canadian prairies include:
* Li, S., MacMillan, R., Lobb, D. A., McConkey, B. G., Moulin, A., and Fraser, W. R.: Lidar DEM error analyses and topographic depression identification in a hummocky landscape in the prairie region of Canada, Geomorphology, 129, 263–275, https://doi.org/10.1016/j.geomorph.2011.02.020, 2011.
* Shaw, D. A., Pietroniro, A., and Martz, L. W.: Topographic analysis for the prairie pothole region of Western Canada, Hydrol. Process., 27, 3105–3114, https://doi.org/10.1002/hyp.9409, 2013.

The Shaw reference seems quite close to your work here.

In addition, while ethically I can by no means suggest that you cite work by your referees here, I would like to point you towards the set of three papers that we have recently written/published; the first one (FlowFill) seems very similar to your approach here. The second and third address the issues of computational efficiency using a graph. Perhaps they would be of more general interest?
* Callaghan, K. L. and Wickert, A. D.: Computing water flow through complex landscapes – Part 1: Incorporating depressions in flow routing using FlowFill, Earth Surf. Dynam., 7, 737–753, https://doi.org/10.5194/esurf-7-737-2019, 2019.
* Barnes, R., Callaghan, K. L., and Wickert, A. D.: Computing water flow through complex landscapes – Part 2: Finding hierarchies in depressions and morphological segmentations, Earth Surf. Dynam., 8, 431–445, https://doi.org/10.5194/esurf-8-431-2020, 2020.
* Barnes, R., Callaghan, K. L., and Wickert, A. D.: Computing water flow through complex landscapes – Part 3: Fill–Spill–Merge: flow routing in depression hierarchies, Earth Surf. Dynam., 9, 105–121, https://doi.org/10.5194/esurf-9-105-2021, 2021.

In addition, our above papers spend some time discussing past work in managing flow through depressions -- indeed, this is where I found the references that I note above.

I hope that this is helpful! Again, and to be explicit: this is not a request to cite the referees' work! But some acknowledgment of prior research in the same vein would, I think, be necessary background for your paper.
