### 2021-04-04 #2276 @KevinShook
Thanks for the suggestions. I was trying to keep the paper to be as short as possible, but will elaborate on other models
WDPM is very different from the work of Shaw et al, which is based on the modelling of simulated individual depressions, rather than a DEM - we have expanded on this methodology in another one of our models, which I was trying to avoid discussing.

WDPM is also unlike programs like FlowMapR, used by Li et al,, and other similar GIS programs, which quantify _filled_ depressions - the intent of WDPM is to quantity the states of water in DEMs which are _not_ completely filled, and to demonstrate the difference in landscape connectivity between filling and emptying fluxes.

The three papers that you have published look to be very interesting, and do seem to be much more similar to our methodology. I am sorry to admit that I was not aware of their existence. I will read them over very carefully.
