### 2021-04-04 #2276 @awickert
In this case, don't feel any need to discuss unrelated literature -- I think that it is simply that I saw that your background included hydrologic and hydraulic models, but I did not see explicit references to prior approaches to manage depressions, which seems to be a key methodological issue that you solve.

I believe that our "FlowFill" is largely the same as your apprpoach, and that our DH+FSM may solve for the desired fields much more efficiently. This is not to detract from the value of WDPM, but is to say that we could perhaps follow up after this review to see what might be helpful in terms of computational efficiency.
