### 2021-04-04 #2276 @KevinShook
Thanks. I'm going to add a brief section differentiating WDPM from other models, which may avoid some confusion.
I'd be very interested in following up offline. WDPM is not very computationally efficient, but it does
have some advantages over stateless models, as its purpose was to model the states of depressional storage
(the spatial distributions of water, and the connected/contributing fractions), under repeated cycles of additions
and removals of water, which few other models do.
