### 2021-04-04 ad32ea7
increased comparison of WDPM with other models
