+++ paper/paper.bib
@@ -1,4 +1,40 @@
 
+@article{ahmedNovelModelStorage2020,
+  title = {A Novel Model for Storage Dynamics Simulation and Inundation Mapping in the Prairies},
+  author = {Ahmed, Mohamed Ismaiel and Elshorbagy, Amin and Pietroniro, Alain},
+  year = {2020},
+  month = nov,
+  volume = {133},
+  pages = {104850},
+  issn = {1364-8152},
+  doi = {10.1016/j.envsoft.2020.104850},
+  url = {https://www.sciencedirect.com/science/article/pii/S1364815220309075},
+  urldate = {2021-04-04},
+  abstract = {The Canadian prairies are dominated by numerous depressions, which can modify the lateral transfer of water to prairie streams. Few studies were conducted to simulate the pothole dynamics using their actual spatial distributions. This study proposes a computationally efficient Prairie Region Inundation MApping (PRIMA) model as a hydrologic routing model, for a more accurate and comprehensive storage dynamics simulation and inundation mapping in the prairies. PRIMA shows potential for simulating the potholes' extents when comparing its results with remote sensing data of pothole areas with an accuracy of 85\% averaged over two prairie basins in Saskatchewan, Canada. PRIMA is three~\textasciitilde ~eight times as computationally efficient as the recently developed Wetland DEM Ponding Model (WDPM). Due to its computational efficiency and ability to provide a good simulation of inundation extents, PRIMA shows strengths as a possible tool for pothole inundation mapping and storage dynamics simulations.},
+  file = {/home/kevin/data/zotero/storage/CCVQLJ5G/S1364815220309075.html},
+  journal = {Environmental Modelling \& Software},
+  keywords = {Contributing area,Distributed modelling,Hysteresis,Inundation mapping,Prairie pothole,Remote sensing},
+  language = {en}
+}
+
+@article{ahmedNovelModelStorage2020a,
+  title = {A Novel Model for Storage Dynamics Simulation and Inundation Mapping in the Prairies},
+  author = {Ahmed, Mohamed Ismaiel and Elshorbagy, Amin and Pietroniro, Alain},
+  year = {2020},
+  month = nov,
+  volume = {133},
+  pages = {104850},
+  issn = {1364-8152},
+  doi = {10.1016/j.envsoft.2020.104850},
+  url = {https://www.sciencedirect.com/science/article/pii/S1364815220309075},
+  urldate = {2021-04-04},
+  abstract = {The Canadian prairies are dominated by numerous depressions, which can modify the lateral transfer of water to prairie streams. Few studies were conducted to simulate the pothole dynamics using their actual spatial distributions. This study proposes a computationally efficient Prairie Region Inundation MApping (PRIMA) model as a hydrologic routing model, for a more accurate and comprehensive storage dynamics simulation and inundation mapping in the prairies. PRIMA shows potential for simulating the potholes' extents when comparing its results with remote sensing data of pothole areas with an accuracy of 85\% averaged over two prairie basins in Saskatchewan, Canada. PRIMA is three~\textasciitilde ~eight times as computationally efficient as the recently developed Wetland DEM Ponding Model (WDPM). Due to its computational efficiency and ability to provide a good simulation of inundation extents, PRIMA shows strengths as a possible tool for pothole inundation mapping and storage dynamics simulations.},
+  file = {/home/kevin/data/zotero/storage/4MWANM8C/S1364815220309075.html},
+  journal = {Environmental Modelling \& Software},
+  keywords = {Contributing area,Distributed modelling,Hysteresis,Inundation mapping,Prairie pothole,Remote sensing},
+  language = {en}
+}
+
 @incollection{armstrongUSINGWETLANDPONDING2013a,
[…]
