+++ paper/paper.bib
@@ -321,7 +321,7 @@
 
 @article{shookMemoryEffectsDepressional2011,
   title = {Memory Effects of Depressional Storage in {{Northern Prairie}} Hydrology},
-  author = {Shook, Kevin R and Pomeroy, John W},
+  author = {Shook, Kevin and Pomeroy, John W},
   year = {2011},
   month = dec,
   volume = {25},
