+++ paper/paper.md
@@ -47,9 +47,9 @@ WDPM does not attempt to compute the magnitudes of the additive or subtractive f
 
 The model output is the depth of water over each cell of the DEM. WDPM is not a hydraulic model; it is a model of the state of the distribution of water in a landscape after the fluxes have been redistributed. FLUXOS-OVERFLOW is a 2-D hydrodynamic model capable of modelling overland flow rates in the prairies [@costaPredictingVariableContributing2020], but it is much slower than WDPM, and it requires more parameters and more forcing data.  
 
-The methodology of WDPM is similar to other simple water distribution models such as the Scalable Pothole terraIn anaLysis aLgorithm (SPILL) [@shawTopographicAnalysisPrairie2012], and FlowFill [callaghanComputingWaterFlow2019] which also apply specified fluxes of water to DEMs. Apart from the differences in the algorithms used by these programs and that of WDPM (which is described below), the greatest difference between WDPM and these models is that WDPM is intended to model the state of water on a Prairie landscape, including its spatial distribution and the connected/contributing fraction of the basin. Thus WDPM is intended to be run using successive additions and removals of water. Although several models allow for landscapes to be filled with water, few simulate its removal.  
+The methodology of WDPM is similar to other simple water distribution models such as the Scalable Pothole terraIn anaLysis aLgorithm (SPILL) [@shawTopographicAnalysisPrairie2012], and FlowFill [@callaghanComputingWaterFlow2019] which also apply specified fluxes of water to DEMs. Apart from the differences in the algorithms used by these programs and that of WDPM (which is described below), the greatest difference between WDPM and these models is that WDPM is intended to model the state of water on a Prairie landscape, including its spatial distribution and the connected/contributing fraction of the basin. Thus WDPM is intended to be run using successive additions and removals of water. Although several models allow for landscapes to be filled with water, few simulate its removal.  
 
-The WDPM also differs from simplified models such as the Pothole Cascade Model (PCM) [@shookStorageDynamicsSimulations2013; @shookQuantifyingEffectsPrairie2020] which applies water to depressions which are modelled as discrete storages. Although such models are very computationally efficient, and can reproduce the variable connected/contributing fractions of Prairie basins [@shookStorageDynamicsSimulations2013], very similarly to WDPM, they cannot simulate the aggregation and disaggregation of water areas, and cannot model the spatial distribution of water in a Prairie basin.  
+The WDPM also differs from simplified models such as the Pothole Cascade Model (PCM) [@shookStorageDynamicsSimulations2013; @shookQuantifyingEffectsPrairie2021] which applies water to depressions which are modelled as discrete storages. Although such models are very computationally efficient, and can reproduce the variable connected/contributing fractions of Prairie basins [@shookStorageDynamicsSimulations2013], very similarly to WDPM, they cannot simulate the aggregation and disaggregation of water areas, and cannot model the spatial distribution of water in a Prairie basin.  
 
 The Prairie Region Inundation MApping (PRIMA) model is similar to WDPM in that it allows both additions and removals of water on Prairie landscape DEMs [@ahmedNovelModelStorage2020] through iterative movement of water. PRIMA has been shown to be more efficient than WDPM version 1.0, in that it moved more water per iteration, although WDPM had reduced computational cost per iteration [@ahmedNovelModelStorage2020]. However, PRIMA has not yet been parallelized, so its execution time cannot be compared to that of WDPM. Furthermore, PRIMA has not yet been compared to the newest version 2.0 of WDPM, which is more efficient than version 1.0.  
 
