### 2021-04-04 #2276 @KevinShook
Have added the additional comparisons as discussed with @awickert.
BTW, you were right about Shaw et al - I was thinking of another one of his models.
I've also added a reference to the PRIMA model.
