### 2021-04-05 #2276 @awickert
Thanks, Kevin! I was starting to doubt my memory of that article. I'm also glad that you note the importance of WDPM simulating water removal. (Actually, @KCallaghan, @r-barnes, and I are wrapping up a manuscript that includes evaporation as well as groundwater... let's talk about this when we connect offline too.)
I'll tick the final box after I finish testing your model on my computer. I'll try to be semi-quick about this!
