### 2021-04-05 #2276 @KevinShook
That's great - I'm looking forward to talking about this.
BTW, I was running WDPM w/o the GUI today and found that the switch to select the CPU/GPU for parallel
processing had been reversed - it should be 0 for the CPU, and 1 for the GPU, but the program was using the reverse. It doesn't affect the GUI, but it's being fixed right now.
