### 2021-04-05 61c9fe4
Update WDPMCL.c
For the opencl parameter, "0" means using CPU and "1" means using GPU
