### 2021-04-05 #2276 @KevinShook
The CPU/GPU switch has been fixed so that it now corresponds to the documentation.
