### 2021-04-18 CentreForHydrology/WDPM#18 @r-barnes
PyLint is still giving bad scores (4.07/10). A lot of this is whitespace which it is nice to have fixed in a consistent way, but which I don't actually care too much about. Ignoring that gives 9.26/10, but reveals a couple of problematic places.

One is in src/WDPM.py:
```
self.lblname5.Enable(True) if lblname5==1 else self.lblname5.Enable(False)
        self.editname5.Enable(True) if editname5==1 else self.editname5.Enable(False)
        self.lblname6.Enable(True) if lblname6==1 else self.lblname6.Enable(False)
        self.editname6.Enable(True) if editname6==1 else self.editname6.Enable(False)
        self.lblname7.Enable(True) if lblname7==1 else self.lblname7.Enable(False)
        self.editname7.Enable(True) if editname7==1 else self.editname7.Enable(False)
        self.lblname5a.Enable(True) if lblname5a==1 else self.lblname5a.Enable(False)
        self.editname5a.Enable(True) if editname5a==1 else self.editname5a.Enable(False)
        self.lblname7a.Enable(True) if lblname7a==1 else self.lblname7a.Enable(False)
        self.editname7a.Enable(True) if editname7a==1 else self.editname7a.Enable(False)
        self.lblname6b.Enable(True) if lblname6b==1 else self.lblname6b.Enable(False)
        self.editname6b.Enable(True) if editname6b==1 else self.editname6b.Enable(False)
        self.lblname7b.Enable(True) if lblname7b==1 else self.lblname7b.Enable(False)
        self.editname7b.Enable(True) if editname7b==1 else self.editname7b.Enable(False)
        self.lblname9a.Enable(True) if lblname9a==1 else self.lblname9a.Enable(False)
        self.txt9a.Enable(True) if txt9a==1 else self.txt9a.Enable(False)
        self.button19a.Enable(True) if button19a==1 else self.button19a.Enable(False)
        self.txt1.Enable(True) if txt1==1 else self.txt1.Enable(False)
        self.button11.Enable(True) if button11==1 else self.button11.Enable(False)
        self.txt2.Enable(True) if txt2==1 else self.txt2.Enable(False)
        self.button12.Enable(True) if button12==1 else self.button12.Enable(False)
        self.txt3.Enable(True) if txt3==1 else self.txt3.Enable(False)
        self.txt4.Enable(True) if txt4==1 else self.txt4.Enable(False)
        self.lblname8.Enable(True) if lblname8==1 else self.lblname8.Enable(False)
        self.combo8.Enable(True) if combo8==1 else self.combo8.Enable(False)
        self.button18.Enable(True) if button18==1 else self.button18.Enable(False)
        self.lblname9.Enable(True) if lblname9==1 else self.lblname9.Enable(False)
```
This is not a good way to write this code! And PyLint complains about it pretty clearly, e.g.,
```
src/WDPM.py:402:8: W0106: Expression "self.lblname9.Enable(True) if lblname9 == 1 else self.lblname9.Enable(False)" is assigned to nothing (expression-not-assigned)
```
The way to write, e.g.,
```
self.lblname9.Enable(True) if lblname9==1 else self.lblname9.Enable(False)
```
is
```
self.lblname9.Enable(lblname9==1)
```
