### 2021-04-18 CentreForHydrology/WDPM#18 @r-barnes
PyLint also complains:
```
src/WDPM.py:713:4: R0912: Too many branches (69/12) (too-many-branches)
```
Looking at this we see several hunks like:
```
        if method=='add':
            demfilename = str(self.txt1.GetValue())
            waterfilename = str(self.txt2.GetValue())
```
The trick here is to make some new methods:
```
Size.run_simulation_optimized_add()
```
And to call these as appropriate from the existing `run_simulation_optimized()`, rather than having a massive ur-function.
