### 2021-04-18 CentreForHydrology/WDPM#19 @r-barnes
I think the `requirements.txt` file should be in more standard location of the root directory, rather than `src/`.
