### 2021-04-18 CentreForHydrology/WDPM#16 @r-barnes
The use of `set(CMAKE_C_FLAGS "-std=c11")` is antequated. Since, irrc, cmake 3.8 or earlier it is better to use
```
target_compile_features(TARGET_NAME PRIVATE cxx_std_11)
```
otherwise you pollute the flag-space.

https://cmake.org/cmake/help/latest/command/target_compile_features.html#command:target_compile_features

But I don't see why you're setting at C++ compilation flag for C code.
