### 2021-04-18 CentreForHydrology/WDPM#16 @r-barnes
Also, please avoid `set(CMAKE_C_COMPILER "gcc")`. `cmake` will select an appropriate compiler or the user can specify with, e.g., `CC=gcc cmake -DCMAKE_BUILD_TYPE=Release ..`
