### 2021-04-18 CentreForHydrology/WDPM#16 @r-barnes
`project(WDPMCL)` should appear once at the top of the CMakeLists.txt file.
