### 2021-04-18 CentreForHydrology/WDPM#16 @r-barnes
Why does
```
	find_package(PythonInterp 3 REQUIRED)
	find_package(PythonLibs 3 REQUIRED)
```
Appear in the `CMakeLists.txt` file? It doesn't appear as though you're using Python in the compilation process.
