+++ src/CMakeLists.txt
@@ -1,38 +1,26 @@
 cmake_minimum_required (VERSION 3.9)
 
+project(WDPMCL)
+
+find_package(OpenCL REQUIRED)
+
 IF (UNIX AND NOT APPLE)
-	MESSAGE(STATUS "current platform : Linux")
-	project(WDPMCL)
-	SET(CMAKE_C_COMPILER "gcc")
-	add_executable(WDPMCL WDPMCL.c)
-	target_link_libraries(WDPMCL OpenCL)	
-	find_package(PythonInterp 3 REQUIRED)
-	find_package(PythonLibs 3 REQUIRED)
+	message(STATUS "current platform : Linux")
+	SET(SRC WDPMCL.c)
 ELSEIF(APPLE)
-	MESSAGE(STATUS "current platform: MacOS")
-	project(WDPMCL)
-	SET(CMAKE_C_COMPILER "gcc")
-	add_executable(WDPMCL WDPMCL.c)
-	target_link_libraries(WDPMCL "-framework OpenCL")
-	find_package(PythonInterp 3 REQUIRED)
-	find_package(PythonLibs 3 REQUIRED)
+	message(STATUS "current platform: MacOS")
+	SET(SRC WDPMCL.c)
 ELSE()
-	MESSAGE(STATUS "curent platform : windows")
-	project(WDPMCL)
-	set(CMAKE_C_COMPILER "gcc")
-	find_package(Python 3)
-	IF(PYTHON_FOUND)
-		message(STATUS "python3 found")
-	ELSE()
-		message(STATUS "python3 not found, please install the python3")
-	ENDIF()
-	ADD_EXECUTABLE(WDPMCL WDPMCL-threshold6.c)
-	message(STATUS "########################################################################")
-	message(STATUS "please make sure you have set the correct OpenCL directory in CMakeLists.txt")
-	message(STATUS "########################################################################")
-	#please set the directory of OpenCL library 
-	target_link_libraries(WDPMCL PUBLIC "C:/OpenCL/OpenCL_SDK/3.0/lib/x86/OpenCL.lib")
-	#please set the directory of OpenCL header file
-	target_include_directories(WDPMCL PUBLIC "C:/OpenCL/OpenCL_SDK/3.0/include")
-	set(CMAKE_C_FLAGS "-std=c11")
+	message(STATUS "curent platform : windows")
+	add_executable(WDPMCL WDPMCL-threshold6.c)
 ENDIF()
+
+add_executable(WDPMCL ${SRC})
+target_include_directories(WDPMCL PRIVATE ${OpenCL_INCLUDE_DIRS})
+target_link_libraries(WDPMCL PRIVATE ${OpenCL_LIBRARIES})
+
+if(NOT CMAKE_CXX_COMPILER_ID STREQUAL "MSVC")
+	target_compile_options(WDPMCL PRIVATE -Wall -pedantic)
+else()
+	target_compile_options(WDPMCL PRIVATE)
+endif()
