### 2021-04-18 CentreForHydrology/WDPM#16 @r-barnes
I've put up a pull request #56 to fix some of these issues.
