### 2021-04-18 CentreForHydrology/WDPM#20 @r-barnes
Compiling with warning flags should be the default, always. #56 adds them to CMakeLists.txt.

On g++ 10.2 I see:
```
../WDPMCL.c:565:9: warning: the address of ‘DEMFileName’ will always evaluate as ‘true’ [-Waddress]
  565 |     if (!DEMFileName){
      |         ^
../WDPMCL.c:1249:25: warning: unused variable ‘local’ [-Wunused-variable]
 1249 |       size_t global[2], local[2];
      |                         ^~~~~
../WDPMCL.c:309:61: warning: unused variable ‘boundaryendcols’ [-Wunused-variable]
  309 |     int *widths, *startcols, *endcols, *boundarystartcols, *boundaryendcols;
      |                                                             ^~~~~~~~~~~~~~~
../WDPMCL.c:309:41: warning: unused variable ‘boundarystartcols’ [-Wunused-variable]
  309 |     int *widths, *startcols, *endcols, *boundarystartcols, *boundaryendcols;
      |                                         ^~~~~~~~~~~~~~~~~
../WDPMCL.c:309:31: warning: unused variable ‘endcols’ [-Wunused-variable]
  309 |     int *widths, *startcols, *endcols, *boundarystartcols, *boundaryendcols;
      |                               ^~~~~~~
../WDPMCL.c:309:19: warning: unused variable ‘startcols’ [-Wunused-variable]
  309 |     int *widths, *startcols, *endcols, *boundarystartcols, *boundaryendcols;
      |                   ^~~~~~~~~
../WDPMCL.c:309:10: warning: unused variable ‘widths’ [-Wunused-variable]
  309 |     int *widths, *startcols, *endcols, *boundarystartcols, *boundaryendcols;
      |          ^~~~~~
../WDPMCL.c:307:16: warning: unused variable ‘boundary’ [-Wunused-variable]
  307 |     int slice, boundary, cpu, gpu;
      |                ^~~~~~~~
../WDPMCL.c:307:9: warning: unused variable ‘slice’ [-Wunused-variable]
  307 |     int slice, boundary, cpu, gpu;
      |         ^~~~~
../WDPMCL.c:306:18: warning: unused variable ‘arg_count’ [-Wunused-variable]
  306 |     int i, j, k, arg_count,basincount, watercount, row, col, iteration_limit;
      |                  ^~~~~~~~~
../WDPMCL.c:305:30: warning: unused variable ‘gracefulExit’ [-Wunused-variable]
  305 |     bool done, isKillSwitch, gracefulExit;
      |                              ^~~~~~~~~~~~
../WDPMCL.c:305:16: warning: variable ‘isKillSwitch’ set but not used [-Wunused-but-set-variable]
  305 |     bool done, isKillSwitch, gracefulExit;
      |                ^~~~~~~~~~~~
../WDPMCL.c:304:43: warning: unused variable ‘drainn’ [-Wunused-variable]
  304 |     double diffdrain, olddrain, mindrain, drainn;
      |                                           ^~~~~~
../WDPMCL.c:296:54: warning: unused variable ‘sliceformat’ [-Wunused-variable]
  296 |     char InParameterFileName[80], WaterFileName[80], sliceformat[80];
      |                                                      ^~~~~~~~~~~
../WDPMCL.c:296:10: warning: unused variable ‘InParameterFileName’ [-Wunused-variable]
  296 |     char InParameterFileName[80], WaterFileName[80], sliceformat[80];
      |          ^~~~~~~~~~~~~~~~~~~
../WDPMCL.c:295:47: warning: unused variable ‘outformat’ [-Wunused-variable]
  295 |     char DEMFileName[80], OutputFileName[80], outformat[80];
      |                                               ^~~~~~~~~
../WDPMCL.c:293:33: warning: unused variable ‘byte3’ [-Wunused-variable]
  293 |     size_t bytes, byte1, byte2, byte3;
      |                                 ^~~~~
../WDPMCL.c:293:26: warning: unused variable ‘byte2’ [-Wunused-variable]
  293 |     size_t bytes, byte1, byte2, byte3;
      |                          ^~~~~
../WDPMCL.c:293:19: warning: unused variable ‘byte1’ [-Wunused-variable]
  293 |     size_t bytes, byte1, byte2, byte3;
      |                   ^~~~~
../WDPMCL.c:291:12: warning: unused variable ‘d_boundaryendcols’ [-Wunused-variable]
  291 |     cl_mem d_boundaryendcols, d_totaldrain;
      |            ^~~~~~~~~~~~~~~~~
../WDPMCL.c:290:58: warning: unused variable ‘d_boundarystartcols’ [-Wunused-variable]
  290 |     cl_mem d_bigwater, d_bigdem, d_startcols, d_endcols, d_boundarystartcols;
      |                                                          ^~~~~~~~~~~~~~~~~~~
../WDPMCL.c:290:47: warning: unused variable ‘d_endcols’ [-Wunused-variable]
  290 |     cl_mem d_bigwater, d_bigdem, d_startcols, d_endcols, d_boundarystartcols;
      |                                               ^~~~~~~~~
../WDPMCL.c:290:34: warning: unused variable ‘d_startcols’ [-Wunused-variable]
  290 |     cl_mem d_bigwater, d_bigdem, d_startcols, d_endcols, d_boundarystartcols;
      |                                  ^~~~~~~~~~~
../WDPMCL.c:288:12: warning: unused variable ‘num_groups’ [-Wunused-variable]
  288 |     cl_int num_groups;
      |            ^~~~~~~~~~
../WDPMCL.c:287:26: warning: unused variable ‘sum_buffer’ [-Wunused-variable]
  287 |     cl_mem input_buffer, sum_buffer;
      |                          ^~~~~~~~~~
../WDPMCL.c:287:12: warning: unused variable ‘input_buffer’ [-Wunused-variable]
  287 |     cl_mem input_buffer, sum_buffer;
      |            ^~~~~~~~~~~~
../WDPMCL.c:284:16: warning: unused variable ‘jj’ [-Wunused-variable]
  284 |     cl_int ii, jj, err, ll;
      |                ^~
../WDPMCL.c:284:12: warning: unused variable ‘ii’ [-Wunused-variable]
  284 |     cl_int ii, jj, err, ll;
      |            ^~
../WDPMCL.c: In function ‘runoffs’:
../WDPMCL.c:1995:40: warning: unused variable ‘cell_water_elev’ [-Wunused-variable]
 1995 |     double ht_diff, center_water_elev, cell_water_elev, flow;
      |                                        ^~~~~~~~~~~~~~~
../WDPMCL.c:1995:21: warning: unused variable ‘center_water_elev’ [-Wunused-variable]
 1995 |     double ht_diff, center_water_elev, cell_water_elev, flow;
      |                     ^~~~~~~~~~~~~~~~~
../WDPMCL.c: In function ‘main’:
../WDPMCL.c:1520:7: warning: ‘drainvol’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1520 |       print_results(initial_vol, final_vol, drainvol, waterfrac, meanwater, draindepth, maxdepth);
      |       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1520:7: warning: ‘draindepth’ may be used uninitialized in this function [-Wmaybe-uninitialized]
../WDPMCL.c:1315:35: warning: ‘olddrain’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1315 |      diffdrain = (fabs(totaldrain - olddrain))*cellarea;
      |                        ~~~~~~~~~~~^~~~~~~~~~
../WDPMCL.c:1520:7: warning: ‘initial_vol’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1520 |       print_results(initial_vol, final_vol, drainvol, waterfrac, meanwater, draindepth, maxdepth);
      |       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1193:19: warning: ‘bytes’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1193 |      d_bigwater = clCreateBuffer(context, CL_MEM_READ_WRITE|CL_MEM_USE_HOST_PTR,bytes,bbigwater,&err);
      |                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1265:12: warning: ‘btotaldrain’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1265 |       err= clEnqueueReadBuffer(queue, d_totaldrain, CL_FALSE, 0, sizeof(double), btotaldrain, 0, NULL, &event);
      |            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1193:19: warning: ‘bbigwater’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1193 |      d_bigwater = clCreateBuffer(context, CL_MEM_READ_WRITE|CL_MEM_USE_HOST_PTR,bytes,bbigwater,&err);
      |                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1186:31: warning: ‘bbigdem’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1186 |       bbigdem[i+j*(numrows+2)]=bigdem[i][j];
      |       ~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~
../WDPMCL.c:1265:12: warning: ‘queue’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1265 |       err= clEnqueueReadBuffer(queue, d_totaldrain, CL_FALSE, 0, sizeof(double), btotaldrain, 0, NULL, &event);
      |            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1246:13: warning: ‘kernel’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1246 |       err = clSetKernelArg(kernel, 7, sizeof(int), (void *)&oj);
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1536:7: warning: ‘program’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1536 |  err= clReleaseProgram(program);
      |       ^~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1193:19: warning: ‘context’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1193 |      d_bigwater = clCreateBuffer(context, CL_MEM_READ_WRITE|CL_MEM_USE_HOST_PTR,bytes,bbigwater,&err);
      |                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:641:13: warning: ‘gpu’ may be used uninitialized in this function [-Wmaybe-uninitialized]
  641 |    device = create_device(gpu);
      |             ^~~~~~~~~~~~~~~~~~
../WDPMCL.c:1180:9: warning: ‘cpu’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1180 |  else if(cpu == 1)
      |         ^
../WDPMCL.c:1405:13: warning: ‘iteration_limit’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1405 |      else if(iteration_limit == 0)
      |             ^
../WDPMCL.c:1108:6: warning: ‘thres’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1108 |    if(bigwater[i][j]<thres)
      |      ^
../WDPMCL.c:777:51: warning: ‘rof’ may be used uninitialized in this function [-Wmaybe-uninitialized]
  777 |                               water[i][j]=addwater*rof;
      |                                           ~~~~~~~~^~~~
../WDPMCL.c:926:74: warning: ‘subtractwater’ may be used uninitialized in this function [-Wmaybe-uninitialized]
  926 |                                               water[i][j]=max(water[i][j]-subtractwater,0);
      |                                                                          ^
../WDPMCL.c:770:42: warning: ‘addwater’ may be used uninitialized in this function [-Wmaybe-uninitialized]
  770 |                               water[i][j]+=addwater;
      |                                          ^~
../WDPMCL.c:1344:46: warning: ‘drain_tolerance’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1344 |          if (max_diff <= elevation_tolerance || diffdrain < drain_tolerance || (k >= iteration_limit)) {
      |              ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../WDPMCL.c:1344:13: warning: ‘elevation_tolerance’ may be used uninitialized in this function [-Wmaybe-uninitialized]
 1344 |          if (max_diff <= elevation_tolerance || diffdrain < drain_tolerance || (k >= iteration_limit)) {
      |             ^
[2/2] Linking C executable WDPMCL
```
Please fix these.
