### 2021-04-18 #2276 @r-barnes
I was able to make another pass today.

Some issues:
* The build system seems complicated versus what's needed. I've put up a PR CentreForHydrology/WDPM#56 that shows how I'd do this. However, I have not been able to test it on Mac/Windows.
* I've made a few other comments regarding code quality.

I think my biggest concern is that so much of the documentation is wrapped up in a PDF file. I don't really see the advantage of this. In particular, I opened the repo again and was immediately lost as to how to install/use the software. It took me 5 clicks and a Ctrl+F to begin to track down this information. Including this information in the README or linking to explanatory markdown seems like a better approach to me. I'm not sure if JOSS has a policy one way or the other on this.
