### 2021-04-18 3c3da1c
Update CMakeLists.txt
correct a typo in the cmake file. We are actually using the same source code for all platforms (Windows, MacOS, and Linux)
