+++ src/CMakeLists.txt
@@ -26,7 +26,7 @@ ELSE()
 	ELSE()
 		message(STATUS "python3 not found, please install the python3")
 	ENDIF()
-	ADD_EXECUTABLE(WDPMCL WDPMCL-threshold6.c)
+	ADD_EXECUTABLE(WDPMCL WDPMCL.c)
 	message(STATUS "########################################################################")
 	message(STATUS "please make sure you have set the correct OpenCL directory in CMakeLists.txt")
 	message(STATUS "########################################################################")
