### 2021-04-19 ece23e5
improved description of OpenCl matrix subdivision
