### 2021-04-19 33a4c96
Update WDPMCL.c
Fixed the warning. But I am using the gcc 10.2, but I do not see the "uninitialize" warning. others are fixed. I will ask him about the warning
