### 2021-04-19 CentreForHydrology/WDPM#20 @leogean3
Hi, I am using gcc-10.2 and using flag -Wall -pedantic to compile. But I do not see the warning about "maybe-uninitialized". I have pulled a request that fix the other warnings. Are you using additional flags that generate the "maybe-uninitialized" warning?
