### 2021-04-23 #2276 @kbarnhart
@awickert thanks for that update and for completing your review! 🎉

@KCallaghan any chance you could provide an update on your availability to complete your review (note, have also emailed).

Finally, regarding @r-barnes question regarding the status of the documentation in a PDF (vs some other format). Thanks for your patience on this, and I see that @KevinShook has made some updates to the Readme.md in the interim.

JOSS doesn't have a policy regarding the format/hosting of the documentation. This comes in part from wanting to provide flexibility to authors in how to handle documentation that meets the criteria listed in the "documentation" section, while respecting the diversity of software types and languages we see. I've handled examples that span from a docx and pdf, to a GitHub Wiki, Jupyter Notebooks, to a sphinx/doxogen approach. Some examples use a hybrid where the documentation may span multiple mediums. When a pdf is used it is definitely a good practice to provide the source (e.g., in this case the lyx file used and the figures, or in other cases the source docx).

I think the key considerations here are:

1. Is it immediately apparent where the documentation elements are located? This may mean pointing the potential user towards one or more resources if documentation elements are located in multiple locations. Documentation should be obtainable without login/registration ([as per submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements)).

2. Does the documentation contain the required information (statement of need, installation instructions, example usage, functionality documentation, automated tests or similar, community guidelines)?

3. Are each of these elements clear and usable (e.g., a non searchable PDF would be unusable)?

Let me know if there is other clarification I can provide.
