### 2021-04-27 CentreForHydrology/WDPM#4 @KCallaghan
Great - compiling from source seems to work ok for me.
