### 2021-04-27 #2276 @KCallaghan
It looks like the authors have handled all of the issues I brought up previously, thank you! I found running and testing the software significantly easier this time around. I feel comfortable ticking off the remaining items in my review. I do have one note:
 - Page 26 of the manual details the fields to enter when running from the command line, but leaves out the 'water runoff fraction', which can lead to confusion (the error message that comes up in the command line shows the correct order of fields)

 And one question:
  - Is there a particular use case in which you think it would be sensible to 'add', and _not_ 'drain'? It seems as though the examples you give require add then drain, and I was wondering why the algorithm initially dams water at the edges of the DEM during the add step rather than allowing water to flow out of the DEM during this step (essentially combining 'add' and 'drain' into one).
