### 2021-04-27 #2276 @KevinShook
Whoops! Thanks for finding the error on P. 26. I'll add it right away.

There are a couple of reasons why you might not want to drain.
One is that in many prairie basins there is no need to drain, as the basin is closed (does not spill)
for any foreseeable addition of water. There's no point in draining these basins; in fact draining
can cause problems, because the program will be trying to remove water from the lowest point,
which would be at the bottom of a depression as the "outlet" lies above much of the basin.

The figure shows one of these basins, where we do a lot of work, near Saskatoon. The figure the effect of adding nearly 2 years of precipitation with 100% runoff and zero evaporation/infiltration, which would be required to fill the basin.
Unsurprisingly, in 50+ years of study, the basin has never spilled. My standard joke is that the last time it spilled was when the glaciers melted!

![600_0_0_0_u](https://user-images.githubusercontent.com/11050573/116308162-6d76c300-a764-11eb-84f3-2ed741393952.png)

Another reason for draining separately is that it might be desirable to use differing precisions when adding/removing water
and draining. When several additions and/or removals are specified as a sequence, the water areas
are required to be accurate, but the water depths are not as important. Therefore, a fairly coarse
tolerance can be used. When draining a basin (particularly for estimating the contributing area),
finer tolerances are usually desirable.
