### 2021-04-27 fab5934
fixed error messages in manual
