### 2021-04-27 #2276 @KevinShook
Glad I could make it clearer. It's such a weird area.
I updated the manual in commit https://github.com/CentreForHydrology/WDPM/commit/fab593467c62d921c92fc16556add57c5a23b0b5
