### 2021-04-28 c9a44e4
Update CMakeLists.txt
Ah sorry, I forget I revoke that request. Here is it, Windows is just using the same source code as other platforms
