### 2021-04-28 c698410
Update WDPMCL.c
show more decimal in the output summary
