### 2021-04-28 #2276 @kbarnhart
@KCallaghan thanks for completing your review and finding that issue. @KevinShook thanks for addressing these issues quickly as they arise and providing the above example. Very interesting.

@r-barnes as you are able please indicate if there are any remaining concerns identified by your review. As always, let me know if I can assist or clarify in the process.
