### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
I've run Pylint again with
```bash
pylint --disable=C0326,C0330 src/
```
and see
```text
src/WDPM.py:392:43: C0303: Trailing whitespace (trailing-whitespace)
src/WDPM.py:711:0: C0303: Trailing whitespace (trailing-whitespace)
src/WDPM.py:958:0: C0303: Trailing whitespace (trailing-whitespace)
src/WDPM.py:1:0: C0302: Too many lines in module (1023/1000) (too-many-lines)
src/WDPM.py:1:0: C0103: Module name "WDPM" doesn't conform to snake_case naming style (invalid-name)
src/WDPM.py:35:0: E0401: Unable to import 'wx' (import-error)
src/WDPM.py:68:19: W0621: Redefining name 'string' from outer scope (line 29) (redefined-outer-name)
src/WDPM.py:63:0: R0903: Too few public methods (1/2) (too-few-public-methods)
src/WDPM.py:72:0: R0903: Too few public methods (1/2) (too-few-public-methods)
src/WDPM.py:88:0: R0902: Too many instance attributes (72/7) (too-many-instance-attributes)
src/WDPM.py:130:8: W0622: Redefining built-in 'quit' (redefined-builtin)
src/WDPM.py:90:4: R0915: Too many statements (143/50) (too-many-statements)
src/WDPM.py:357:4: E0202: An attribute defined in src.WDPM line 649 hides this method (method-hidden)
src/WDPM.py:371:4: R0913: Too many arguments (28/5) (too-many-arguments)
src/WDPM.py:371:4: R0914: Too many local variables (28/15) (too-many-locals)
src/WDPM.py:742:4: C0116: Missing function or method docstring (missing-function-docstring)
src/WDPM.py:742:4: R0914: Too many local variables (18/15) (too-many-locals)
src/WDPM.py:747:8: W0612: Unused variable 'checkpointfilenamex' (unused-variable)
src/WDPM.py:742:4: R0912: Too many branches (21/12) (too-many-branches)
src/WDPM.py:742:4: R0915: Too many statements (69/50) (too-many-statements)
src/WDPM.py:819:4: C0116: Missing function or method docstring (missing-function-docstring)
src/WDPM.py:819:4: R0914: Too many local variables (17/15) (too-many-locals)
src/WDPM.py:824:8: W0612: Unused variable 'checkpointfilenamex' (unused-variable)
src/WDPM.py:819:4: R0912: Too many branches (20/12) (too-many-branches)
src/WDPM.py:819:4: R0915: Too many statements (65/50) (too-many-statements)
src/WDPM.py:890:4: C0116: Missing function or method docstring (missing-function-docstring)
src/WDPM.py:890:4: R0914: Too many local variables (17/15) (too-many-locals)
src/WDPM.py:895:8: W0612: Unused variable 'checkpointfilenamex' (unused-variable)
src/WDPM.py:890:4: R0912: Too many branches (20/12) (too-many-branches)
src/WDPM.py:890:4: R0915: Too many statements (64/50) (too-many-statements)
src/WDPM.py:88:0: R0904: Too many public methods (41/20) (too-many-public-methods)
```
I've mentioned fixes for the "Too many branches" problem above, so I'm surprised to see that it remains unresolved.

It would also be good if you could drop the unused variables (if they are actually unused).

The "Too many public methods" problem is dealt with by prefixing class method names with an underscore ("_") to indicate that they should only be used internally to the class and not externally by downstream users. That is in
```
class MyClass:
  def foo(self):
    ...
  def run(self):
    ...
```
both `run` and `foo` are publicly accessible while in
```
class MyClass:
  def _foo(self):
    ...
  def run(self):
    self._foo()
```
only `run` should be used by someone other than you.
