### 2021-05-19 CentreForHydrology/WDPM#20 @r-barnes
I've tried again and don't see the maybe-uninitialized warning any more on either gcc-10.2 or gcc-9.3, so I think we're good here.
