### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
This section (and similar)
```
        if demfilename=='':
            self.on_error_dem()
            plat='error'
        if waterfilename=='':
            self.on_error_water()
            plat='error'
        if wateroutputfilename=='':
            self.on_error_output()
            plat='error'
        if checkpointfilename=='':
            self.on_error_check_point()
            plat='error'
```
could be simplified, eg, by introducing a function:
```
def _error_callback_on_empty(plat, s, callback):
  if s=="":
    callback()
    return True
  return False

errored = False
errored |= self._error_callback_on_empty(demfilename, self.on_error_dem)
errored |= self._error_callback_on_empty(waterfilename, self.on_error_water)
errored |= self._error_callback_on_empty(wateroutputfilename, self.on_error_output)
errored |= self._error_callback_on_empty(checkpointfilename, self.on_error_check_point)
```
