### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
While something like this:
```
        if method1=="Serial CPU":
            method1="0"
        elif method1=="OpenCL":
            method1="1"
```
can be written:
```
methods = {"Serial CPU": "0", "OpenCL": "1"} # Put this somewhere relatively global
methods.get(method1, None) # Use this somewhere local
```
Note that this makes it easy to extend the code.
