### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
These sections are mostly the same:
```
        if plat in ('Darwin', 'Linux'):
            cmd = [solver, method, demfilename, waterfilename, wateroutputfilename,
                                        checkpointfilename, waterdepths,
                                         elevationtol, method1, method2,threshold, limitation]
            self.module2(cmd)
        else:
            cmd = [solverw, method, demfilename, waterfilename, wateroutputfilename,
                                        checkpointfilename, waterdepths,
                                         elevationtol, method1, method2,threshold, limitation]
```
So why not:
```
        cmd = [method, demfilename, waterfilename, wateroutputfilename,
               checkpointfilename, waterdepths,
               elevationtol, method1, method2, threshold, limitation]
        if plat in ('Darwin', 'Linux'):
            self.module2(cmd + [solver])
        else:
            self.module2(cmd + [solverw])
```
