### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
I think
```
raise "Unable to remove file: report.txt" from AssertionError
```
is invalid syntax
