### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
PyLint is correct that this is too many parameters to the function. This is asking for problems if you add or remove them.
```
    def en_dis_control(self, lblname5, editname5, lblname6, editname6, lblname7, editname7,
    		lblname5a, editname5a, lblname7a, editname7a, lblname6b, editname6b, lblname7b,
    		editname7b, lblname9a, txt9a, button19a, txt1, button11, txt2, button12, txt3,
    		txt4, lblname8, combo8, button18, lblname9):
```
Instead, it is better to use a parameter packing technique like a dictionary:
```
# Somewhere global-ish
add_panels_enabled_disabled = {
  "lblname5": True,
  "editname5": False,
  ...
}

# In class
def en_dis_control(self, enabled_disabled):
  self.lblname5.Enable(enabled_disabled["lblname5"])
  ...

# Alternatively you can shrink the function dramatically by relying on `enabled_disabled` to specify attribute names
def en_dis_control(self, enabled_disabled):
  for k, v in enabled_disabled.items():
    getattr(self, k).Enable(v)

# Either way, later
        if method=='add':
            self.Bind(wx.EVT_MENU, self.en_dis_control(add_panels_enabled_disabled))
```
