### 2021-05-19 CentreForHydrology/WDPM#18 @r-barnes
```
self.dirname5: str = str()
```
can be written
```
self.dirname5: str = ""
```
