### 2021-05-19 CentreForHydrology/WDPM#67 @r-barnes
Tests
I'm not finding a test suite.

[This site](https://realpython.com/pytest-python-testing/) explains how to build one.

In our recent FillSpillMerge work we used randomized testing to generate hundreds of thousands of artificial landscapes and then tested
1. To see if adding a very large amount of water gives a result exactly equal to depression filling,
2. That incrementally adding portions of water gives the same result as adding the same amount of water all at once
3. And more, see [link](https://github.com/r-barnes/Barnes2020-FillSpillMerge/blob/master/unittests/fsm_tests.cpp) for details.

Performing extensive tests identified bugs in our software and we were able to fix them before they impacted downstream science. Given the complexity of WDPM having a good set of unittests seems appropriate.
