### 2021-05-19 CentreForHydrology/WDPM#68 @r-barnes
Contributors guide
I can't find a contributors guide.

See https://github.com/holukas/dyco/blob/main/CONTRIBUTING.md for an example
