### 2021-05-19 CentreForHydrology/WDPM#69 @r-barnes
API Documentation
"Computation Settings" don't appear to be documented in the manual. Other settings from the GUI may also be undocumented.

Although the manual should document everything since you're otherwise relying heavily on it for this purpose, using tooltips in the GUI would be a much more accessible way of communicating the information, eg, https://codeloop.org/how-to-create-tooltip-in-wxpython/
