### 2021-05-19 CentreForHydrology/WDPM#70 @r-barnes
C code formatting
In the C code I see inconsistent formatting like so:
```
    final_vol=final_vol*cellarea;
    meanwater = watertotal/((float)watercount);
    basincount=0;
      for(i=0; i<numrows; i++){
    for(j=0; j<numcols; j++){
	    if(dem[i][j]>missingvalue){
            basincount++;
	    }
	}
    }
```
This is because the code intermixes tabs and spaces. It's likely that my tab display width is different from yours.

Please reformat the code to consistently use only spaces for whitespace (this is a common requirement of coding [style guides](https://google.github.io/styleguide/cppguide.html#Spaces_vs._Tabs)).
