### 2021-05-19 #2276 @r-barnes
Okay, I've made another review. There are a few things left to fix.

* [ ] **Tests.** The software doesn't seem to have tests associated with it. The manual obliquely refers to a sample `basin5.asc` file which could be processed; however, this would only show that the software does something. Automating this simple test of functionality with unittest+pytest should be pretty straight-forward. However, more extensive tests are desirable. For instance, my recent FillSpillMerge work performs a very similar operation to what the authors describe and during testing we discovered several subtle bugs. After this recent scarring experience I've been trying to more thoroughly test all my own code! In FillSpillMerge I used randomized testing to generate hundreds of thousands of artificial landscapes and then tested (1) to see if adding a very large amount of water gives a result exactly equal to depression filling, (2) that incrementally adding portions of water gives the same result as adding the same amount of water all at once, and (3) more tests as described [here](https://github.com/r-barnes/Barnes2020-FillSpillMerge/blob/master/unittests/fsm_tests.cpp). I'm glad that my tests allowed me to identify and fix bugs in my code before they impacted downstream science. Given the complexity of WDPM having a good set of unittests seems appropriate. CentreForHydrology/WDPM#67
* [ ] **Contributing.** I couldn't find a clear contributor guide. I expected one in a `CONTRIBUTING.md` file but don't see that or "Contributing" in the manual. Regardless of whether this material is present in the manual, I think it should be in a `CONTRIBUTING.md` file because this is, logically, where a contributor would look for it. CentreForHydrology/WDPM#88
* [ ] **Functionality documentation.** Certain GUI elements seem to be undocumented. Using Tooltips to provide documentation information in-program seems like a reasonable solution. CentreForHydrology/WDPM#69
* [ ] **Functionality.** As above, I can confirm that the software does something. However, I cannot confirm that it does exactly what it claims it does. A test suite such as that discussed above would make a strong argument in favour of functionality.
* [ ] **State of the Field**. My FillSpillMerge software, mentioned above and described in [two](https://esurf.copernicus.org/articles/8/431/2020/) [papers](https://esurf.copernicus.org/articles/9/105/2021/), seems to perform a very similar operation to what the authors describe here. Admittedly, I have a bias/conflict of interest here, so I'm not interested in pressing this issue if the authors prefer not to mention it.
* [ ] **Code Quality**. This is much improved from the previous pass. CentreForHydrology/WDPM#18 and CentreForHydrology/WDPM#70 identify a few further improvements that I think would improve readability and modularity.
