+++ CONTRIBUTING.md
@@ -1,0 +1,27 @@
+## How to contribute
+Anyone interested is encouraged to contribute to the repository by **forking and submitting a pull request**. 
+We require that anyone contributing to this project adhere to the [code of conduct](https://github.com/CentreForHydrology/WDPM/CODE_OF_CONDUCT.md). Please read it before commenting or contributing code.  
+
+If you are new to GitHub, you might start with a [basic tutorial](https://help.github.com/articles/set-up-git) and  check out a more detailed guide to [pull requests](https://help.github.com/articles/using-pull-requests/).
+
+
+
+### Feature request
+
+* Submit an issue on the [Issues page](https://github.com/CentreForHydrology/WDPM/issues) or;
+* Submit a clean pull request with your code contribution for review. Pull requests will be evaluated by the repository guardians on a schedule and if deemed beneficial will be committed to the master.
+
+### Code contributions
+
+* Fork this repo to your Github account
+* Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/[your username]/WDPM.git`
+* Make sure to track progress upstream (i.e., on our version of `WDPM` at `CentreForHydrology/WDPM`) by doing `git remote add upstream https://github.com/CentreForHydrology/WDPM.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
+* Make your changes (bonus points for making changes on a new feature branch)
+* Push up to your account
+* Submit a pull request to home base (likely master branch, but check to make sure) at `CentreForHydrology/WDPM`
+
+### Prefer to Email?
+
+You can email kevin.shook@usask.ca with comments or questions.
+
+### Thanks for contributing!
