+++ CODE_OF_CONDUCT.md
@@ -34,8 +34,8 @@ Conduct may be permanently removed from the project team.
 This Code of Conduct applies both within project spaces and in public spaces
 when an individual is representing the project or its community.
 
-Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting a project maintainer at sam.albers@gov.bc.ca. All complaints will be reviewed and investigated 
+Instances of abusive, harassing, or otherwise unacceptable behaviour may be
+reported by contacting a project maintainer at kevin.shook@usask.ca. All complaints will be reviewed and investigated 
 and will result in a response that is deemed necessary and appropriate to the 
 circumstances. Maintainers are obligated to maintain confidentiality with regard 
 to the reporter of an incident.
