### 2021-05-19 CentreForHydrology/WDPM#68 @KevinShook
Thanks for pointing this out.
We are such a small project, that I never considered it.
We have added CONTRUIBUTING.md as well as a code of conduct in commit
https://github.com/CentreForHydrology/WDPM/commit/09d5787a938a751dff4b68837a0753d582090c0a
