### 2021-05-19 b27e344
updated manual to mention title 'Computation Settings'
