### 2021-05-19 CentreForHydrology/WDPM#69 @KevinShook
"Computation Settings" is a title, not a program setting.
All of the individual settings were documented in the manual, however the title
is now referred to explicitly in the manual, in commit
https://github.com/CentreForHydrology/WDPM/commit/b27e3446d9f78b83c58e8a0d4f0bebd608060eb5
