### 2021-06-16 CentreForHydrology/WDPM#67 @KevinShook
Thanks for the suggestion. We are looking into setting up automated unit tests.
The OpenCL kernel file `runoff.cl`, which does the parallelised water distribution calculations, is very small (183 lines)
and has not changed since 2014 (other than the very recent change from tabs to spaces),
so we hadn't been using automated tests.

Our focus has always been using on WDPM for real-world landscapes. Because WDPM outputs standard .asc files, we
have used GIS scripts in R to validate the results of WDPM against prior versions, for differing landscapes.
As we are interested in the connected/contributing fractions of prairie landscapes, and how they change with the state of water, it's been important to evaluate any changes in the program outputs with respect to many sequences of additions and removals of water.
