### 2021-06-21 8930b08
Update WDPM.py
fix several pylint warning. there are two left. Maybe I need to ask for further advice:)
