+++ src/WDPM.py
@@ -65,9 +65,9 @@ class RedirectText:
     def __init__(self,aWxTextCtrl):
         self.out=aWxTextCtrl
 
-    def write(self,string):
+    def write(self,strings):
         '''write'''
-        wx.CallAfter(self.out.WriteText, string)
+        wx.CallAfter(self.out.WriteText, strings)
 
 class CharValidator(wx.PyValidator):
     '''check if the char is validate'''
@@ -127,9 +127,9 @@ class Size(wx.Frame):
         self.flagz = 0
         menubar = wx.MenuBar()
         filez = wx.Menu()
-        quit = wx.MenuItem(filez, 1, '&Quit\tCtrl+Q')
+        quits = wx.MenuItem(filez, 1, '&Quit\tCtrl+Q')
         about = wx.MenuItem(filez, 3, '&About\tCtrl+A')
-        filez.Append(quit)
+        filez.Append(quits)
         filez.Append(about)
         self.Bind(wx.EVT_MENU, self.on_quit, id=1)
         self.Bind(wx.EVT_MENU, self.on_about, id=3)
@@ -238,20 +238,20 @@ class Size(wx.Frame):
         self.log = wx.TextCtrl(self.panel, -1, pos=(16*x_p, x_p), size=(25*x_p, 22*x_p),
                                        style = wx.TE_MULTILINE|wx.TE_READONLY)
         ##pre-define
-        self.dirname5: str = str()
-        self.dirname2: str = str()
-        self.filename2: str = str()
-        self.filename1: str = str()
-        self.dirname1: str = str()
-        self.dirname0: str = str()
-        self.filename0: str = str()
-        self.threadx: str = str()
-        self.processx: str = str()
-        self.dirname5x: str = str()
-        self.filename5: str = str()
-        self.repx: str = str()
-        self.thread1: str = str()
-        self.rep2: str = str()
+        self.dirname5: str = ""
+        self.dirname2: str = ""
+        self.filename2: str = ""
+        self.filename1: str = ""
+        self.dirname1: str = ""
+        self.dirname0: str = ""
+        self.filename0: str = ""
+        self.threadx: str = ""
+        self.processx: str = ""
+        self.dirname5x: str = ""
+        self.filename5: str = ""
+        self.repx: str = ""
+        self.thread1: str = ""
+        self.rep2: str = ""
 
         font1 = wx.Font(11, wx.FONTFAMILY_DEFAULT, wx.FONTSTYLE_NORMAL,
         		wx.FONTWEIGHT_NORMAL, False, 'Consolas')
@@ -737,34 +737,29 @@ class Size(wx.Frame):
                     cmd = [solverw, filename]
                     self.module2(cmd)
 
+    def _error_callback_on_empty(self, sent, callback):
+        if sent=="":
+            callback()
+            return True
+        return False
+
     def check_file(self):
         '''check the parameter'''
-        method1 = self.combo8.GetValue()
-        method2 = self.combo9.GetValue()
         demfilename = str(self.txt1.GetValue())
         waterfilename = str(self.txt2.GetValue())
         wateroutputfilename = os.path.join(self.txt0a.GetValue(),str(self.txt3.GetValue()))
         checkpointfilename = os.path.join(self.txt0a.GetValue(),"temp.asc")
-        elevationtol = str(self.editname7.GetValue())
         if self.txt4.GetValue()=='NULL':
             checkpointfilename = str(self.txt4.GetValue())
         else:
             checkpointfilename = os.path.join(self.txt0a.GetValue(),str(self.txt4.GetValue()))
-        if demfilename=='':
-            self.on_error_dem()
-        if waterfilename=='':
-            self.on_error_water()
-        if wateroutputfilename=='':
-            self.on_error_output()
[…]
