### 2021-06-28 #2276 @KevinShook
Thanks for getting back.
We have responded to issues https://github.com/CentreForHydrology/WDPM/issues/70  and https://github.com/CentreForHydrology/WDPM/issues/69 and
https://github.com/CentreForHydrology/WDPM/issues/68 and
