### 2021-06-29 CentreForHydrology/WDPM#67 @kbarnhart
@KevinShook sorry I didn't notice your comment here until now.

One point I'd like to make is that JOSS is less concerned with how testing is implemented over whether appropriate tests have been developed and there are instructions on how to run and evaluate them. This is to say, if a user is asked to run a script from the command line which runs a series of additions and removals (perhaps on your basin5.asc file) and then compares the results to a set of benchmark files, that is totally fine. While an automatic test suite is certainly the "best in class" version of testing, it is not necessarily required. The more important characteristic of testing is that an appropriately designed set of tests exist.

I hope this point is helpful. Please let me know if I might be able to provide additional clarifications or guidance.
