### 2021-06-30 CentreForHydrology/WDPM#67 @KevinShook
As the primary output from WDPM is just a file of water depths, it requires quite a bit of post-processing. For example, these are the types of figures that I typically use. The first shows frequency distributions of water areas for 3 sequences (addition, removal and subsequent addition of water). It is necessary to find all of the water areas in each model output dataset, which requires GIS or an R script.

![Figure10](https://user-images.githubusercontent.com/11050573/123988154-030d1980-d985-11eb-84a1-55b0ba00fd0e.png)

The second shows the change in the connected fractions of a single basin, using DEMs with varying resolutions. In this case, each point is computed by adding (or removing) water, draining, then adding 1 mm, followed by draining. The change in the water storage, when subtracted from the volume of water added, and divided by the same volume is used to represent the connected/contributing fraction. We also use an R script to do this.
![WDPMresolutionScalingROF](https://user-images.githubusercontent.com/11050573/123987654-9bef6500-d984-11eb-940c-88e0c8aa2ab9.png)

However, I very much agree with your point, that it is important that the user be able to validate that the model is running correctly. I think that we can do this by using the WDPM log files. Normally these are just used to ensure that the
program is running correctly. However, the log file also contains several statistics w.r.t. the state of the water
storage at the end of the run, such as the total water volume, mean/max depth and areal water fraction.

I will set up scripts to add, subtract and drain water, and will explain their use, and the results that should be
expected by the user. Would that be OK?
