### 2021-06-30 CentreForHydrology/WDPM#67 @kbarnhart
@KevinShook I think there are a lot of different ways to go about constructing tests for this type of algorithm. One way could be the approach you describe. I'm not quite sure that this approach is fully necessary, though.

An approach I've found useful for constructing test is to identify a synthetic case that (a) is complex enough to interact with the full algorithm and (b) is small enough that I can hand calculate the expected outcome from the algorithm. For example, I might use something like the below DEM. Constructed so it has two depressions, one larger and more complex, and one smaller.

```
10.0    10.0    10.0    10.0   10.0    10.0
10.0     9.0     8.0     9.0   10.0    10.0
10.0     7.0     5.0    10.0   10.0    10.0
10.0    10.0    10.0    10.0    4.0    10.0
10.0    10.0    10.0    10.0   10.0    10.0
```

I might then perform a series of add/subtract/drain and at each step assert that the results are correct. Such an assertion could could manifest in a few ways. For example, it could assert that mass is conserved and that the spatial distribution of water depths is as expected, both within the expected tolerance.

Such an approach tests the core algorithms, but on such a small scale that it is possible to know what the results should be.

There are a few ways any sort of testing approach might be implemented. One option would be to write a script that sets up these tests (could include writing a temporary .asc file with the synthetic grid), runs them, prints a useful message ("All tests pass", statement about where tests failed) based on the comparison between model results and expected results, and then cleans up any temporary files. Another option would be to use packages like pytest which help with discovery and evaluation of tests.

Hope this is helpful guidance. Happy to continue discussing as needed.
