### 2021-07-01 CentreForHydrology/WDPM#67 @KevinShook
Thanks. I had been thinking of something like this, but was stumped by the need to build
a system which would not require the development of further software (which would itself
require testing), and would be platform-agnostic. Something like this can easily be done in Linux
(which I use exclusively), probably using awk, but is much more difficult for Windows, which is used by
the majority of people running WDPM. However, with the advent of WSL we should be able to build a single
script.
