### 2021-07-09 24af034
changes to validation instructions
