### 2021-07-09 CentreForHydrology/WDPM#67 @KevinShook
I have added scripts to validate the WDPM output. They are in the `/validation` directory.
The scripts are described in the file `WDPM_validation.md`.
There's a bash script which runs WDPM using the add, drain and subtract modules. It also
calls 3 awk scripts which test the outputs. All 3 scripts check the water stored in a small depression.
The `add_test.awk` scripts also independently checks the conservation of mass. The `drain_test.awk` script checks that
the remaining volume of water is correct, and also checks that the depth at the outlet is zero.
The `subtract_test.awk` script also checks the remaining volume of water.
