### 2021-08-04 #2276 @kbarnhart
@KevinShook after discussion with others on the JOSS editorial board, I have decided to move forward with this submission.

Many thanks to all three reviewers @r-barnes @awickert and @KCallaghan for your constructive reviews. I believe that the authors have now addressed the outstanding comments, and I have gone ahead and completed the remaining checklist items/closed issues.

The checklist items I checked from @r-barnes 's checklist were:
Functionality: Have the functional claims of the software been confirmed?
Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
 State of the field: Do the authors describe how this software compares to other commonly-used packages?
