### 2021-08-04 #2276 @kbarnhart
@KevinShook a few final comments from me:
1. In the paper, please fix the .bib entry for both Venema references, I think the curly braces should be removed so the author is listed as Venema, H.D. (or correct me if you think otherwise).
2. Over the last year we have started enforcing that the papers actually contain a section called "Statement of Need". I would recommend you rename "Program Description" to "Statement of Need" and then add another section title around line 39 to demarcate the transition between the prior section and your nice comparison between this and other work.

Once you have done both of those things, please verify that `@whedon check references` and `@whedon generate pdf` work without error (e.g., that all DOIs are resolved and the PDF successfully compiles).  At that point, please do the following:

 - [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).

I will be out of the office on Friday August 6th
