### 2021-08-05 df73c6d
added section Statement of Need and updated a reference
