+++ paper/paper.bib
@@ -428,7 +428,7 @@
 
 @techreport{venemahenrydavidClimateAdaptationUsing2020,
   title = {Climate {{Adaptation Using Natural Infrastructure Town}} of {{Virden}} \& the {{Rural Municipality}} of {{Wallace}} - {{Woodworth}}},
-  author = {{Venema, Henry David}},
+  author = {Venema, Henry David},
   year = {2020},
   month = mar,
   pages = {45},
@@ -439,7 +439,7 @@
 
 @techreport{venemahenrydavidFloodRiskMapping2020,
   title = {Flood {{Risk Mapping Study Rural Municipality}} of {{Springfield}}},
-  author = {{Venema, Henry David}},
+  author = {Venema, Henry David},
   year = {2020},
   month = jun,
   pages = {105},
