### 2021-08-05 #2276 @KevinShook
Should I be storing the paper on Zenodo as well? As it is under a CC4 licence, I would assume that it would have to be kept separate from the program which is GPL3 - is that correct?
