### 2021-08-05 #2276 @kbarnhart
The entire repo at https://github.com/CentreForHydrology/WDPM is what should be in the Zenodo archive.

I'd expect that the .bib and the .md for the paper are covered under the repository's license, while the final publication is what would be under the CC4. I am not a license expert though.

I think that answers the question, but if it doesn't, let me know.
