### 2021-08-05 #2276 @KevinShook
The tagged release (v2.0b) is now on GitHub:
https://github.com/CentreForHydrology/WDPM/releases/tag/v2.0b
